Module winit::changelog::v0_12

Available on nowhere.

0.12.0

Added subclass to macos windows so they can be made resizable even with no decorations.

Dead keys now work properly on X11, no longer resulting in a panic.

On X11, input method creation first tries to use the value from the user’s XMODIFIERS environment variable, so application developers should no longer need to manually call XSetLocaleModifiers. If that fails, fallbacks are tried, which should prevent input method initialization from ever outright failing.

Fixed thread safety issues with input methods on X11.

Add support for Touch for win32 backend.

Fixed Window::get_inner_size and friends to return the size in pixels instead of points when using HIDPI displays on OSX.