winit::changelog

Module v0_18

Available nowhere.

0.18.1

On macOS, fix Yen (JIS) so applications receive the event.

On X11 with a tiling WM, fixed high CPU usage when moving windows across monitors.

On X11, fixed panic caused by dropping the window before running the event loop.

on macOS, added WindowExt::set_simple_fullscreen which does not require a separate space

Introduce WindowBuilderExt::with_app_id to allow setting the application ID on Wayland.

On Windows, catch panics in event loop child thread and forward them to the parent thread. This prevents an invocation of undefined behavior due to unwinding into foreign code.

On Windows, fix issue where resizing or moving window combined with grabbing the cursor would freeze program.

On Windows, fix issue where resizing or moving window would eat Awakened events.

On Windows, exiting fullscreen after entering fullscreen with disabled decorations no longer shrinks window.

On X11, fixed a segfault when using virtual monitors with XRandR.

Derive Ord and PartialOrd for VirtualKeyCode enum.

On Windows, fix issue where hovering or dropping a non file item would create a panic.

On Wayland, fix resizing and DPI calculation when a wl_output is removed without sending a leave event to the wl_surface, such as disconnecting a monitor from a laptop.

On Wayland, DPI calculation is handled by smithay-client-toolkit.

On X11, WindowBuilder::with_min_dimensions and WindowBuilder::with_max_dimensions now correctly account for DPI.

Added support for generating dummy DeviceIds and WindowIds to better support unit testing.

On macOS, fixed unsoundness in drag-and-drop that could result in drops being rejected.

On macOS, implemented WindowEvent::Refresh.

On macOS, all MouseCursor variants are now implemented and the cursor will no longer reset after unfocusing.

Removed minimum supported Rust version guarantee.

0.18.0

Breaking: image crate upgraded to 0.20. This is exposed as part of the icon_loading API.

On Wayland, pointer events will now provide the current modifiers state.

On Wayland, titles will now be displayed in the window header decoration.

On Wayland, key repetition is now ended when keyboard loses focus.

On Wayland, windows will now use more stylish and modern client side decorations.

On Wayland, windows will use server-side decorations when available.

Breaking: Added support for F16-F24 keys (variants were added to the VirtualKeyCode enum).

Fixed graphical glitches when resizing on Wayland.

On Windows, fix freezes when performing certain actions after a window resize has been triggered. Reintroduces some visual artifacts when resizing.

Updated window manager hints under X11 to v1.5 of Extended Window Manager Hints.

Added WindowBuilderExt::with_gtk_theme_variant to X11-specific WindowBuilder functions.

Fixed UTF8 handling bug in X11 set_title function.

On Windows, Window::set_cursor now applies immediately instead of requiring specific events to occur first.

On Windows, the HoveredFile and HoveredFileCancelled events are now implemented.

On Windows, fix Window::set_maximized.

On Windows 10, fix transparency (#260).

On macOS, fix modifiers during key repeat.

Implemented the Debug trait for Window, EventsLoop, EventsLoopProxy and WindowBuilder.

On X11, now a Resized event will always be generated after a DPI change to ensure the window’s logical size is consistent with the new DPI.

Added further clarifications to the DPI docs.

On Linux, if neither X11 nor Wayland manage to initialize, the corresponding panic now consists of a single line only.

Add optional serde feature with implementations of Serialize/Deserialize for DPI types and various event types.

Add PartialEq, Eq, and Hash implementations on public types that could have them but were missing them.

On X11, drag-and-drop receiving an unsupported drop type can no longer cause the WM to freeze.

Fix issue whereby the OpenGL context would not appear at startup on macOS Mojave (#1069).

Breaking: Removed From<NSApplicationActivationPolicy> impl from ActivationPolicy on macOS.

On macOS, the application can request the user’s attention with WindowExt::request_user_attention.