Module winit::changelog::v0_19

Available on nowhere.

0.19.1

On Wayland, added a get_wayland_display function to EventsLoopExt.

On Windows, fix CursorMoved(0, 0) getting dispatched on window focus.

On macOS, fix command key event left and right reverse.

On FreeBSD, NetBSD, and OpenBSD, fix build of X11 backend.

On Linux, the numpad’s add, subtract and divide keys are now mapped to the Add, Subtract and Divide virtual key codes

On macOS, the numpad’s subtract key has been added to the Subtract mapping

On Wayland, the numpad’s home, end, page up and page down keys are now mapped to the Home, End, PageUp and PageDown virtual key codes

On Windows, fix icon not showing up in corner of window.

On X11, change DPI scaling factor behavior. First, winit tries to read it from “Xft.dpi” XResource, and uses DPI calculation from xrandr dimensions as fallback behavior.

0.19.0

On X11, we will use the faster XRRGetScreenResourcesCurrent function instead of XRRGetScreenResources when available.

On macOS, fix keycodes being incorrect when using a non-US keyboard layout.

On Wayland, fix with_title() not setting the windows title

On Wayland, add set_wayland_theme() to control client decoration color theme

Added serde serialization to os::unix::XWindowType.

Breaking: Remove the icon_loading feature and the associated image dependency.

On X11, make event loop thread safe by replacing XNextEvent with select(2) and XCheckIfEvent

On Windows, fix malformed function pointer typecast that could invoke undefined behavior.

Refactored Windows state/flag-setting code.

On Windows, hiding the cursor no longer hides the cursor for all Winit windows - just the one hide_cursor was called on.

On Windows, cursor grabs used to get perpetually canceled when the grabbing window lost focus. Now, cursor grabs automatically get re-initialized when the window regains focus and the mouse moves over the client area.

On Windows, only vertical mouse wheel events were handled. Now, horizontal mouse wheel events are also handled.

On Windows, ignore the AltGr key when populating the ModifiersState type.