Module winit::changelog::v0_21

Available on nowhere.

0.21.0

On Windows, fixed “error: linking with link.exe failed: exit code: 1120” error on older versions of windows.

On macOS, fix set_minimized(true) works only with decorations.

On macOS, add hide_application to EventLoopWindowTarget via a new EventLoopWindowTargetExtMacOS trait. hide_application will hide the entire application by calling -[NSApplication hide: nil].

On macOS, fix not sending ReceivedCharacter event for specific keys combinations.

On macOS, fix CursorMoved event reporting the cursor position using logical coordinates.

On macOS, fix issue where unbundled applications would sometimes open without being focused.

On macOS, fix run_return does not return unless it receives a message.

On Windows, fix bug where RedrawRequested would only get emitted every other iteration of the event loop.

On X11, fix deadlock on window state when handling certain window events.

WindowBuilder now implements Default.

Breaking: WindowEvent::CursorMoved changed to f64 units, preserving high-precision data supplied by most backends

On Wayland, fix coordinates in mouse events when scale factor isn’t 1

On Web, add the ability to provide a custom canvas

Breaking: On Wayland, the WaylandTheme struct has been replaced with a Theme trait, allowing for extra configuration