Module winit::changelog::v0_24

Available on nowhere.

0.24.0

On Windows, fix applications not exiting gracefully due to thread_event_target_callback accessing corrupted memory.

On Windows, implement Window::set_ime_position.

Breaking: On Windows, Renamed WindowBuilderExtWindows’s is_dark_mode to theme.

Breaking: On Windows, renamed WindowBuilderExtWindows::is_dark_mode to theme.

On Windows, add WindowBuilderExtWindows::with_theme to set a preferred theme.

On Windows, fix bug causing message boxes to appear delayed.

On Android, calling WindowEvent::Focused now works properly instead of always returning false.

On Windows, fix Alt-Tab behaviour by removing borderless fullscreen “always on top” flag.

On Windows, fix bug preventing windows with transparency enabled from having fully-opaque regions.

Breaking: On Windows, include prefix byte in scancodes.

On Wayland, fix window not being resizeable when using WindowBuilder::with_min_inner_size.

On Unix, fix cross-compiling to wasm32 without enabling X11 or Wayland.

On Windows, fix use-after-free crash during window destruction.

On Web, fix WindowEvent::ReceivedCharacter never being sent on key input.

On macOS, fix compilation when targeting aarch64.

On X11, fix Window::request_redraw not waking the event loop.

On Wayland, the keypad arrow keys are now recognized.

Breaking Rename desktop::EventLoopExtDesktop to run_return::EventLoopExtRunReturn.

Added request_user_attention method to Window.

Breaking: On macOS, removed WindowExt::request_user_attention, use Window::request_user_attention.

Breaking: On X11, removed WindowExt::set_urgent, use Window::request_user_attention.

On Wayland, default font size in CSD increased from 11 to 17.

On Windows, fix bug causing message boxes to appear delayed.

On Android, support multi-touch.

On Wayland, extra mouse buttons are not dropped anymore.

Breaking: MouseButton::Other now uses u16.