Module winit::changelog::v0_25

Available on nowhere.

0.25.0

Breaking: On macOS, replace WindowBuilderExtMacOS::with_activation_policy with EventLoopExtMacOS::set_activation_policy

On macOS, wait with activating the application until the application has initialized.

On macOS, fix creating new windows when the application has a main menu.

On Windows, fix fractional deltas for mouse wheel device events.

On macOS, fix segmentation fault after dropping the main window.

On Android, InputEvent::KeyEvent is partially implemented providing the key scancode.

Added is_maximized method to Window.

On Windows, fix bug where clicking the decoration bar would make the cursor blink.

On Windows, fix bug causing newly created windows to erroneously display the “wait” (spinning) cursor.

On macOS, wake up the event loop immediately when a redraw is requested.

On Windows, change the default window size (1024x768) to match the default on other desktop platforms (800x600).

On Windows, fix bug causing mouse capture to not be released.

On Windows, fix fullscreen not preserving minimized/maximized state.

On Android, unimplemented events are marked as unhandled on the native event loop.

On Windows, added WindowBuilderExtWindows::with_menu to set a custom menu at window creation time.

On Android, bump ndk and ndk-glue to 0.3: use predefined constants for event ident.

On macOS, fix objects captured by the event loop closure not being dropped on panic.

On Windows, fixed WindowEvent::ThemeChanged not properly firing and fixed Window::theme returning the wrong theme.

On Web, added support for DeviceEvent::MouseMotion to listen for relative mouse movements.

Added WindowBuilder::with_position to allow setting the position of a Window on creation. Supported on Windows, macOS and X11.

Added Window::drag_window. Implemented on Windows, macOS, X11 and Wayland.

On X11, bump mio to 0.7.

On Windows, added WindowBuilderExtWindows::with_owner_window to allow creating popup windows.

On Windows, added WindowExtWindows::set_enable to allow creating modal popup windows.

On macOS, emit RedrawRequested events immediately while the window is being resized.

Implement Default, Hash, and Eq for LogicalPosition, PhysicalPosition, LogicalSize, and PhysicalSize.

On macOS, initialize the Menu Bar with minimal defaults. (Can be prevented using enable_default_menu_creation)

On macOS, change the default behavior for first click when the window was unfocused. Now the window becomes focused and then emits a MouseInput event on a “first mouse click”.

Implement mint (math interoperability standard types) conversions (under feature flag mint).