winit::changelog

Module v0_26

Available nowhere.

0.26.1

Fix linking to the ColorSync framework on macOS 10.7, and in newer Rust versions.

On Web, implement cursor grabbing through the pointer lock API.

On X11, add mappings for numpad comma, numpad enter, numlock and pause.

On macOS, fix Pinyin IME input by reverting a change that intended to improve IME.

On Windows, fix a crash with transparent windows on Windows 11.

0.26.0

Update raw-window-handle to v0.4. This is not a breaking change, we still implement HasRawWindowHandle from v0.3, see rust-windowing/raw-window-handle#74. Note that you might have to run cargo update -p raw-window-handle after upgrading.

On X11, bump mio to 0.8.

On Android, fixed WindowExtAndroid::config initially returning an empty Configuration.

On Android, fixed Window::scale_factor and MonitorHandle::scale_factor initially always returning 1.0.

On X11, select an appropriate visual for transparency if is requested

On Wayland and X11, fix diagonal window resize cursor orientation.

On macOS, drop the event callback before exiting.

On Android, implement Window::request_redraw

Breaking: On Web, remove the stdweb backend.

Added Window::focus_windowto bring the window to the front and set input focus.

On Wayland and X11, implement is_maximized method on Window.

On Windows, prevent ghost window from showing up in the taskbar after either several hours of use or restarting explorer.exe.

On macOS, fix issue where ReceivedCharacter was not being emitted during some key repeat events.

On Wayland, load cursor icons hand2 and hand1 for CursorIcon::Hand.

Breaking: On Wayland, Theme trait and its support types are dropped.

On Wayland, bump smithay-client-toolkit to 0.15.1.

On Wayland, implement request_user_attention with xdg_activation_v1.

On X11, emit missing WindowEvent::ScaleFactorChanged when the only monitor gets reconnected.

On X11, if RANDR based scale factor is higher than 20 reset it to 1

On Wayland, add an enabled-by-default feature called wayland-dlopen so users can opt out of using dlopen to load system libraries.

Breaking: On Android, bump ndk and ndk-glue to 0.5.

On Windows, increase wait timer resolution for more accurate timing when using WaitUntil.

On macOS, fix native file dialogs hanging the event loop.

On Wayland, implement a workaround for wrong configure size when using xdg_decoration in kwin_wayland

On macOS, fix an issue that prevented the menu bar from showing in borderless fullscreen mode.

On X11, EINTR while polling for events no longer causes a panic. Instead it will be treated as a spurious wakeup.