Enum winit::WindowEvent

pub enum WindowEvent {
    Resized(u32, u32),
    Moved(i32, i32),
    Closed,
    DroppedFile(PathBuf),
    HoveredFile(PathBuf),
    HoveredFileCancelled,
    ReceivedCharacter(char),
    Focused(bool),
    KeyboardInput {
        device_id: DeviceId,
        input: KeyboardInput,
    },
    MouseMoved {
        device_id: DeviceId,
        position: (f64, f64),
    },
    MouseEntered {
        device_id: DeviceId,
    },
    MouseLeft {
        device_id: DeviceId,
    },
    MouseWheel {
        device_id: DeviceId,
        delta: MouseScrollDelta,
        phase: TouchPhase,
    },
    MouseInput {
        device_id: DeviceId,
        state: ElementState,
        button: MouseButton,
    },
    TouchpadPressure {
        device_id: DeviceId,
        pressure: f32,
        stage: i64,
    },
    AxisMotion {
        device_id: DeviceId,
        axis: AxisId,
        value: f64,
    },
    Refresh,
    Suspended(bool),
    Touch(Touch),
}

Variants
Resized(u32, u32)

The size of the window has changed.

Moved(i32, i32)

The position of the window has changed.

Closed

The window has been closed.

DroppedFile(PathBuf)

A file has been dropped into the window.

HoveredFile(PathBuf)

A file is being hovered over the window.

HoveredFileCancelled

A file was hovered, but has exited the window.

ReceivedCharacter(char)

The window received a unicode character.

Focused(bool)

The window gained or lost focus.

The parameter is true if the window has gained focus, and false if it has lost focus.

KeyboardInput

An event from the keyboard has been received.

Fields of KeyboardInput

device_id: DeviceId

input: KeyboardInput

MouseMoved

The cursor has moved on the window.

position is (x,y) coords in pixels relative to the top-left corner of the window. Because the range of this data is limited by the display area and it may have been transformed by the OS to implement effects such as mouse acceleration, it should not be used to implement non-cursor-like interactions such as 3D camera control.

Fields of MouseMoved

device_id: DeviceId

position: (f64, f64)

MouseEntered

The cursor has entered the window.

Fields of MouseEntered

device_id: DeviceId

MouseLeft

The cursor has left the window.

Fields of MouseLeft

device_id: DeviceId

MouseWheel

A mouse wheel movement or touchpad scroll occurred.

Fields of MouseWheel

device_id: DeviceId

delta: MouseScrollDelta

phase: TouchPhase

MouseInput

An mouse button press has been received.

Fields of MouseInput

device_id: DeviceId

state: ElementState

button: MouseButton

TouchpadPressure

Touchpad pressure event.

At the moment, only supported on Apple forcetouch-capable macbooks. The parameters are: pressure level (value between 0 and 1 representing how hard the touchpad is being pressed) and stage (integer representing the click level).

Fields of TouchpadPressure

device_id: DeviceId

pressure: f32

stage: i64

AxisMotion

Motion on some analog axis not otherwise handled. May overlap with mouse motion.

Fields of AxisMotion

device_id: DeviceId

axis: AxisId

value: f64

Refresh

The window needs to be redrawn.

Suspended(bool)

App has been suspended or resumed.

The parameter is true if app was suspended, and false if it has been resumed.

Touch(Touch)

Touch event has been received