winit
Module error

Structs

NotSupportedError

The error type for when the requested operation is not supported by the backend.

OsError

The error type for when the OS cannot perform the requested operation.

Enums

EventLoopError

A general error that may occur while running the Winit event loop

ExternalError

An error that may be generated when requesting Winit state