Module winit::event::device

Raw hardware events that are not associated with any particular window.

Useful for interactions that diverge significantly from a conventional 2D GUI, such as 3D camera or first-person game controls. Many physical actions, such as mouse movement, can produce both device and window events. Because window events typically arise from virtual devices (corresponding to GUI cursors and keyboard focus) the device IDs may not match.

All attached devices are guaranteed to emit an Added event upon the initialization of the event loop.

Note that device events are always delivered regardless of window focus.

Structs

GamepadHandle
A handle to a gamepad/joystick device.

KeyboardId
A typed identifier for a keyboard device.

MouseId
A typed identifier for a mouse device.

Enums

GamepadAxis
A hint suggesting the type of axis that moved.

GamepadButton
A hint suggesting the type of button that was pressed.

GamepadEvent
Gamepad/joystick events.

KeyboardEvent
Raw keyboard events.

MouseEvent
Raw mouse events.

RumbleError
Error reported if a rumble attempt unexpectedly failed.

Side
A given joystick's side on the gamepad.