Enum winit::event::device::GamepadEvent

pub enum GamepadEvent {
    Added,
    Removed,
    Axis {
        axis_id: AxisId,
        axis: Option<GamepadAxis>,
        value: f64,
        stick: bool,
    },
    Stick {
        x_id: AxisId,
        y_id: AxisId,
        x_value: f64,
        y_value: f64,
        side: Side,
    },
    Button {
        button_id: ButtonId,
        button: Option<GamepadButton>,
        state: ElementState,
    },
}

Gamepad/joystick events.

These can be generated by any of a variety of game controllers, including (but not limited to) gamepads, joysicks, and HOTAS devices.

Variants
Added

A gamepad/joystick device has been added.

Removed

A gamepad/joystick device has been removed.

Axis

An analog axis value on the gamepad/joystick has changed.

Winit does NOT provide deadzone filtering, and such filtering may have to be provided by API users for joystick axes.

Fields of Axis

axis_id: AxisIdaxis: Option<GamepadAxis>

A hint regarding the physical axis that moved.

On traditional gamepads (such as an X360 controller) this can be assumed to have a non-None value; however, other joystick devices with more varied layouts generally won't provide a value here.

TODO: DISCUSS CONTROLLER MAPPING ONCE WE FIGURE OUT WHAT WE'RE DOING THERE.

value: f64stick: bool

Whether or not this axis has also produced a GamepadEvent::Stick event.

Stick

A two-axis joystick's value has changed.

This is mainly provided to assist with deadzone calculation, as proper deadzones should be calculated via the combined distance of each joystick axis from the center of the joystick, rather than per-axis.

Note that this is only guaranteed to be emitted for traditionally laid out gamepads. More complex joysticks generally don't report specifics of their layout to the operating system, preventing Winit from automatically aggregating their axis input into two-axis stick events.

Fields of Stick

x_id: AxisId

The X axis' ID.

y_id: AxisId

The Y axis' ID.

x_value: f64y_value: f64side: Side

Which joystick side produced this event.

Button
Fields of Button

button_id: ButtonIdbutton: Option<GamepadButton>

A hint regarding the location of the button.

The caveats on the Axis.hint field also apply here.

state: ElementState