winit::event_loop

Struct ActiveEventLoop

pub struct ActiveEventLoop { /* private fields */ }

Target that associates windows with an EventLoop.

This type exists to allow you to create new windows while Winit executes your callback.

Implementations

impl ActiveEventLoop

pub fn create_window( &self, window_attributes: WindowAttributes, ) -> Result<Window, OsError>

Create the window.

Possible causes of error include denied permission, incompatible system, and lack of memory.

Platform-specific

Web: The window is created but not inserted into the web page automatically. Please see the web platform module for more information.

pub fn create_custom_cursor( &self, custom_cursor: CustomCursorSource, ) -> CustomCursor

Create custom cursor.

pub fn available_monitors(&self) -> impl Iterator<Item = MonitorHandle>

Returns the list of all the monitors available on the system.

pub fn primary_monitor(&self) -> Option<MonitorHandle>

Returns the primary monitor of the system.

Returns None if it canโt identify any monitor as a primary one.

Platform-specific

Wayland / Web: Always returns None.

pub fn listen_device_events(&self, allowed: DeviceEvents)

Change if or when DeviceEvents are captured.

Since the DeviceEvent capture can lead to high CPU usage for unfocused windows, winit will ignore them by default for unfocused windows on Linux/BSD. This method allows changing this at runtime to explicitly capture them again.

Platform-specific

Wayland / macOS / iOS / Android / Orbital: Unsupported.

pub fn system_theme(&self) -> Option<Theme>

Returns the current system theme.

Returns None if it cannot be determined on the current platform.

Platform-specific

iOS / Android / Wayland / x11 / Orbital: Unsupported.

pub fn set_control_flow(&self, control_flow: ControlFlow)

Sets the ControlFlow.

pub fn control_flow(&self) -> ControlFlow

Gets the current ControlFlow.

pub fn exit(&self)

This exits the event loop.

See LoopExiting.

pub fn exiting(&self) -> bool

Returns if the EventLoop is about to stop.

See exit().

pub fn owned_display_handle(&self) -> OwnedDisplayHandle

Gets a persistent reference to the underlying platform display.

See the OwnedDisplayHandle type for more information.