Struct winit::event_loop::EventLoopBuilder

pub struct EventLoopBuilder<T: 'static> { /* private fields */ }

Object that allows building the event loop.

This is used to make specifying options that affect the whole application easier. But note that constructing multiple event loops is not supported.

Implementations

impl EventLoopBuilder<()>

pub fn new() -> Self

Start building a new event loop.

impl<T> EventLoopBuilder<T>

pub fn with_user_event() -> Self

Start building a new event loop, with the given type as the user event type.

pub fn build(&mut self) -> Result<EventLoop<T>, EventLoopError>

Builds a new event loop.

For cross-platform compatibility, the EventLoop must be created on the main thread, and only once per application.

Calling this function will result in display backend initialisation.

Panics

Attempting to create the event loop off the main thread will panic. This restriction isn’t strictly necessary on all platforms, but is imposed to eliminate any nasty surprises when porting to platforms that require it. EventLoopBuilderExt::any_thread functions are exposed in the relevant platform module if the target platform supports creating an event loop on any thread.

Platform-specific

Wayland/X11: to prevent running under Wayland or X11 unset WAYLAND_DISPLAY or DISPLAY respectively when building the event loop.

Android: must be configured with an AndroidApp from android_main() by calling .with_android_app(app) before calling .build(), otherwise it’ll panic.