Struct winit::event_loop::EventLoopWindowTarget

pub struct EventLoopWindowTarget<T: 'static> { /* private fields */ }

Target that associates windows with an EventLoop.

This type exists to allow you to create new windows while Winit executes your callback. EventLoop will coerce into this type (impl<T> Deref for EventLoop<T>), so functions that take this as a parameter can also take &EventLoop.

Implementations

impl<T> EventLoopWindowTarget<T>

pub fn available_monitors(&self) -> impl Iterator<Item = MonitorHandle>

Returns the list of all the monitors available on the system.

pub fn primary_monitor(&self) -> Option<MonitorHandle>

Returns the primary monitor of the system.

Returns None if it can’t identify any monitor as a primary one.

Platform-specific

Wayland / Web: Always returns None.

pub fn listen_device_events(&self, allowed: DeviceEvents)

Change if or when DeviceEvents are captured.

Since the DeviceEvent capture can lead to high CPU usage for unfocused windows, winit will ignore them by default for unfocused windows on Linux/BSD. This method allows changing this at runtime to explicitly capture them again.

Platform-specific

Wayland / macOS / iOS / Android / Orbital: Unsupported.

pub fn set_control_flow(&self, control_flow: ControlFlow)

Sets the ControlFlow.

pub fn control_flow(&self) -> ControlFlow

Gets the current ControlFlow.

pub fn exit(&self)

This exits the event loop.

See LoopExiting.

pub fn exiting(&self) -> bool

Returns if the EventLoop is about to stop.

See exit().