Module winit::events

Structs

KeyboardInput
Describes a keyboard input event.

ModifiersState
Represents the current state of the keyboard modifiers

Touch
Represents touch event

Enums

DeviceEvent
Represents raw hardware events that are not associated with any particular window.

ElementState
Describes the input state of a key.

Event
Describes a generic event.

MouseButton
Describes a button of a mouse controller.

MouseScrollDelta
Describes a difference in the mouse scroll wheel state.

TouchPhase
Describes touch-screen input state.

VirtualKeyCode
Symbolic name for a keyboard key.

WindowEvent
Describes an event from a Window.

Type Definitions

AxisId
Identifier for a specific analog axis on some device.

ButtonId
Identifier for a specific button on some device.

ScanCode
Hardware-dependent keyboard scan code.