Enum winit::keyboard::NativeKey

pub enum NativeKey {
    Unidentified,
    Android(u32),
    MacOS(u16),
    Windows(u16),
    Xkb(u32),
    Web(SmolStr),
}

Contains the platform-native logical key identifier

Exactly what that means differs from platform to platform, but the values are to some degree tied to the currently active keyboard layout. The same key on the same keyboard may also report different values on different platforms, which is one of the reasons this is a per-platform enum.

This enum is primarily used to store raw keysym when Winit doesn’t map a given native logical key identifier to a meaningful Key variant. This lets you use Key, and let the user define keybinds which work in the presence of identifiers we haven’t mapped for you yet.

Variants

Unidentified

Android(u32)

An Android “keycode”, which is similar to a “virtual-key code” on Windows.

MacOS(u16)

A macOS “scancode”. There does not appear to be any direct analogue to either keysyms or “virtual-key” codes in macOS, so we report the scancode instead.

Windows(u16)

A Windows “virtual-key code”.

Xkb(u32)

An XKB “keysym”.

Web(SmolStr)

A “key value string”.