winit::monitor

Struct MonitorHandle

pub struct MonitorHandle { /* private fields */ }

Handle to a monitor.

Allows you to retrieve information about a given monitor and can be used in Window creation.

Implementations

impl MonitorHandle

pub fn name(&self) -> Option<String>

Returns a human-readable name of the monitor.

Returns None if the monitor doesnโt exist anymore.

pub fn size(&self) -> PhysicalSize<u32>

Returns the monitorโs resolution.

pub fn position(&self) -> PhysicalPosition<i32>

Returns the top-left corner position of the monitor relative to the larger full screen area.

pub fn refresh_rate_millihertz(&self) -> Option<u32>

The monitor refresh rate used by the system.

Return Some if succeed, or None if failed, which usually happens when the monitor the window is on is removed.

When using exclusive fullscreen, the refresh rate of the VideoModeHandle that was used to enter fullscreen should be used instead.

pub fn scale_factor(&self) -> f64

Returns the scale factor of the underlying monitor. To map logical pixels to physical pixels and vice versa, use Window::scale_factor.

See the dpi module for more information.

Platform-specific

X11: Can be overridden using the WINIT_X11_SCALE_FACTOR environment variable.

Wayland: May differ from Window::scale_factor.

Android: Always returns 1.0.

pub fn video_modes(&self) -> impl Iterator<Item = VideoModeHandle>

Returns all fullscreen video modes supported by this monitor.

Platform-specific

Web: Always returns an empty iterator