Trait winit::os::unix::WindowExt

pub trait WindowExt {
    fn get_xlib_window(&self) -> Option<c_ulong>;

    fn get_xlib_display(&self) -> Option<*mut c_void>;

    fn get_xlib_screen_id(&self) -> Option<c_int>;

    fn get_xlib_xconnection(&self) -> Option<Arc<XConnection>>;

    fn send_xim_spot(&self, x: i16, y: i16);

    fn get_xcb_connection(&self) -> Option<*mut c_void>;

    fn get_wayland_surface(&self) -> Option<*mut c_void>;

    fn get_wayland_display(&self) -> Option<*mut c_void>;

    fn is_ready(&self) -> bool;
}

Additional methods on Window that are specific to Unix.

Required methods

fn get_xlib_window(&self) -> Option<c_ulong>

Returns the ID of the Window xlib object that is used by this window.

Returns None if the window doesn't use xlib (if it uses wayland for example).

fn get_xlib_display(&self) -> Option<*mut c_void>

Returns a pointer to the Display object of xlib that is used by this window.

Returns None if the window doesn't use xlib (if it uses wayland for example).

The pointer will become invalid when the glutin Window is destroyed.

fn get_xlib_screen_id(&self) -> Option<c_int>

fn get_xlib_xconnection(&self) -> Option<Arc<XConnection>>

fn send_xim_spot(&self, x: i16, y: i16)

fn get_xcb_connection(&self) -> Option<*mut c_void>

This function returns the underlying xcb_connection_t of an xlib Display.

Returns None if the window doesn't use xlib (if it uses wayland for example).

The pointer will become invalid when the glutin Window is destroyed.

fn get_wayland_surface(&self) -> Option<*mut c_void>

Returns a pointer to the wl_surface object of wayland that is used by this window.

Returns None if the window doesn't use wayland (if it uses xlib for example).

The pointer will become invalid when the glutin Window is destroyed.

fn get_wayland_display(&self) -> Option<*mut c_void>

Returns a pointer to the wl_display object of wayland that is used by this window.

Returns None if the window doesn't use wayland (if it uses xlib for example).

The pointer will become invalid when the glutin Window is destroyed.

fn is_ready(&self) -> bool

Deprecated

Check if the window is ready for drawing

It is a remnant of a previous implementation detail for the wayland backend, and is no longer relevant.

Always return true.

Loading content...