Trait winit::platform::android::EventLoopBuilderExtAndroid

pub trait EventLoopBuilderExtAndroid {
    // Required methods
    fn with_android_app(&mut self, app: AndroidApp) -> &mut Self;
    fn handle_volume_keys(&mut self) -> &mut Self;
}

Available on android_platform only.

Required Methods

fn with_android_app(&mut self, app: AndroidApp) -> &mut Self

Associates the [AndroidApp] that was passed to android_main() with the event loop

This must be called on Android since the [AndroidApp] is not global state.

fn handle_volume_keys(&mut self) -> &mut Self

Calling this will mark the volume keys to be manually handled by the application

Default is to let the operating system handle the volume keys

Object Safety

This trait is not object safe.