winit::platform::macos

Trait ActiveEventLoopExtMacOS

pub trait ActiveEventLoopExtMacOS {
    // Required methods
    fn hide_application(&self);
    fn hide_other_applications(&self);
    fn set_allows_automatic_window_tabbing(&self, enabled: bool);
    fn allows_automatic_window_tabbing(&self) -> bool;
}

Available on macos_platform only.

Additional methods on ActiveEventLoop that are specific to macOS.

Required Methods

fn hide_application(&self)

Hide the entire application. In most applications this is typically triggered with Command-H.

fn hide_other_applications(&self)

Hide the other applications. In most applications this is typically triggered with Command+Option-H.

fn set_allows_automatic_window_tabbing(&self, enabled: bool)

Set whether the system can automatically organize windows into tabs.

https://developer.apple.com/documentation/appkit/nswindow/1646657-allowsautomaticwindowtabbing

fn allows_automatic_window_tabbing(&self) -> bool

Returns whether the system can automatically organize windows into tabs.