Trait winit::platform::macos::EventLoopWindowTargetExtMacOS

pub trait EventLoopWindowTargetExtMacOS {
    fn hide_application(&self);
    fn hide_other_applications(&self);
}

Additional methods on EventLoopWindowTarget that are specific to macOS.

Required Methods

fn hide_application(&self)

Hide the entire application. In most applications this is typically triggered with Command-H.

fn hide_other_applications(&self)

Hide the other applications. In most applications this is typically triggered with Command+Option-H.