Trait winit::platform::macos::WindowAttributesExtMacOS

pub trait WindowAttributesExtMacOS {
    // Required methods
    fn with_movable_by_window_background(
        self,
        movable_by_window_background: bool,
    ) -> Self;
    fn with_titlebar_transparent(self, titlebar_transparent: bool) -> Self;
    fn with_title_hidden(self, title_hidden: bool) -> Self;
    fn with_titlebar_hidden(self, titlebar_hidden: bool) -> Self;
    fn with_titlebar_buttons_hidden(self, titlebar_buttons_hidden: bool) -> Self;
    fn with_fullsize_content_view(self, fullsize_content_view: bool) -> Self;
    fn with_disallow_hidpi(self, disallow_hidpi: bool) -> Self;
    fn with_has_shadow(self, has_shadow: bool) -> Self;
    fn with_accepts_first_mouse(self, accepts_first_mouse: bool) -> Self;
    fn with_tabbing_identifier(self, identifier: &str) -> Self;
    fn with_option_as_alt(self, option_as_alt: OptionAsAlt) -> Self;
}

Available on macos_platform only.

Additional methods on WindowAttributes that are specific to MacOS.

Note: Properties dealing with the titlebar will be overwritten by the WindowAttributes::with_decorations method:

with_titlebar_transparent

with_title_hidden

with_titlebar_hidden

with_titlebar_buttons_hidden

with_fullsize_content_view

Required Methods

fn with_movable_by_window_background( self, movable_by_window_background: bool, ) -> Self

Enables click-and-drag behavior for the entire window, not just the titlebar.

fn with_titlebar_transparent(self, titlebar_transparent: bool) -> Self

Makes the titlebar transparent and allows the content to appear behind it.

fn with_title_hidden(self, title_hidden: bool) -> Self

Hides the window title.

fn with_titlebar_hidden(self, titlebar_hidden: bool) -> Self

Hides the window titlebar.

fn with_titlebar_buttons_hidden(self, titlebar_buttons_hidden: bool) -> Self

Hides the window titlebar buttons.

fn with_fullsize_content_view(self, fullsize_content_view: bool) -> Self

Makes the window content appear behind the titlebar.

fn with_disallow_hidpi(self, disallow_hidpi: bool) -> Self

fn with_has_shadow(self, has_shadow: bool) -> Self

fn with_accepts_first_mouse(self, accepts_first_mouse: bool) -> Self

Window accepts click-through mouse events.

fn with_tabbing_identifier(self, identifier: &str) -> Self

Defines the window tabbing identifier.

https://developer.apple.com/documentation/appkit/nswindow/1644704-tabbingidentifier

fn with_option_as_alt(self, option_as_alt: OptionAsAlt) -> Self

Set how the Option keys are interpreted.

See WindowExtMacOS::set_option_as_alt for details on what this means if set.

Object Safety

This trait is not object safe.