winit::platform::macos
Trait WindowBuilderExtMacOS

pub trait WindowBuilderExtMacOS {
    // Required methods
    fn with_movable_by_window_background(
        self,
        movable_by_window_background: bool,
    ) -> WindowBuilder;
    fn with_titlebar_transparent(
        self,
        titlebar_transparent: bool,
    ) -> WindowBuilder;
    fn with_title_hidden(self, title_hidden: bool) -> WindowBuilder;
    fn with_titlebar_hidden(self, titlebar_hidden: bool) -> WindowBuilder;
    fn with_titlebar_buttons_hidden(
        self,
        titlebar_buttons_hidden: bool,
    ) -> WindowBuilder;
    fn with_fullsize_content_view(
        self,
        fullsize_content_view: bool,
    ) -> WindowBuilder;
    fn with_disallow_hidpi(self, disallow_hidpi: bool) -> WindowBuilder;
    fn with_has_shadow(self, has_shadow: bool) -> WindowBuilder;
    fn with_accepts_first_mouse(
        self,
        accepts_first_mouse: bool,
    ) -> WindowBuilder;
    fn with_option_as_alt(self, option_as_alt: OptionAsAlt) -> WindowBuilder;
}

Additional methods on WindowBuilder that are specific to MacOS.

Note: Properties dealing with the titlebar will be overwritten by the WindowBuilder::with_decorations method:

with_titlebar_transparent

with_title_hidden

with_titlebar_hidden

with_titlebar_buttons_hidden

with_fullsize_content_view

Required Methods

fn with_movable_by_window_background( self, movable_by_window_background: bool, ) -> WindowBuilder

Enables click-and-drag behavior for the entire window, not just the titlebar.

fn with_titlebar_transparent(self, titlebar_transparent: bool) -> WindowBuilder

Makes the titlebar transparent and allows the content to appear behind it.

fn with_title_hidden(self, title_hidden: bool) -> WindowBuilder

Hides the window title.

fn with_titlebar_hidden(self, titlebar_hidden: bool) -> WindowBuilder

Hides the window titlebar.

fn with_titlebar_buttons_hidden( self, titlebar_buttons_hidden: bool, ) -> WindowBuilder

Hides the window titlebar buttons.

fn with_fullsize_content_view( self, fullsize_content_view: bool, ) -> WindowBuilder

Makes the window content appear behind the titlebar.

fn with_disallow_hidpi(self, disallow_hidpi: bool) -> WindowBuilder

fn with_has_shadow(self, has_shadow: bool) -> WindowBuilder

fn with_accepts_first_mouse(self, accepts_first_mouse: bool) -> WindowBuilder

Window accepts click-through mouse events.

fn with_option_as_alt(self, option_as_alt: OptionAsAlt) -> WindowBuilder

Set whether the OptionAsAlt key is interpreted as the Alt modifier.

See WindowExtMacOS::set_option_as_alt for details on what this means if set.