Module winit::platform::platform::x11

Re-exports

pub use self::monitor::MonitorId;

pub use self::window::UnownedWindow;

pub use self::xdisplay::XConnection;

pub use self::xdisplay::XNotSupported;

pub use self::xdisplay::XError;

Modules

dnd

events

ffi

ime

monitor

util

window

xdisplay

Structs

Device

DeviceId

DeviceInfo

EventsLoop

EventsLoopProxy

GenericEventCookie
XEvents of type GenericEvent store their actual data in an XGenericEventCookie data structure. This is a wrapper to extract the cookie from a GenericEvent XEvent and release the cookie data once it has been processed

ScrollAxis

Window

WindowId

XExtension

Enums

ScrollOrientation

Functions

mkdid

mkwid