Module winit::platform::startup_notify

Available on x11_platform or wayland_platform only.

Window startup notification to handle window raising.

The ActivationToken is essential to ensure that your newly created window will obtain the focus, otherwise the user could be requered to click on the window.

Such token is usually delivered via the environment variable and could be read from it with the EventLoopExtStartupNotify::read_token_from_env.

Such token must also be reset after reading it from your environment with reset_activation_token_env otherwise child processes could inherit it.

When starting a new child process with a newly obtained ActivationToken from WindowExtStartupNotify::request_activation_token the set_activation_token_env must be used to propagate it to the child

To ensure the delivery of such token by other processes to you, the user should set StartupNotify=true inside the .desktop file of their application.

The specification could be found here.

Traits

EventLoopExtStartupNotify

WindowBuilderExtStartupNotify

WindowExtStartupNotify

Functions

reset_activation_token_env

Remove the activation environment variables from the current process.

set_activation_token_env

Set environment variables responsible for activation token.