Enum winit::platform::unix::XWindowType

pub enum XWindowType {

    Desktop,
    Dock,
    Toolbar,
    Menu,
    Utility,
    Splash,
    Dialog,
    DropdownMenu,
    PopupMenu,
    Tooltip,
    Notification,
    Combo,
    Dnd,
    Normal,

}

X window type. Maps directly to _NET_WM_WINDOW_TYPE.

Variants

Desktop

A desktop feature. This can include a single window containing desktop icons with the same dimensions as the screen, allowing the desktop environment to have full control of the desktop, without the need for proxying root window clicks.

Dock

A dock or panel feature. Typically a Window Manager would keep such windows on top of all other windows.

Toolbar

Toolbar windows. “Torn off” from the main application.

Menu

Pinnable menu windows. “Torn off” from the main application.

Utility

A small persistent utility window, such as a palette or toolbox.

Splash

The window is a splash screen displayed as an application is starting up.

Dialog

This is a dialog window.

DropdownMenu

A dropdown menu that usually appears when the user clicks on an item in a menu bar. This property is typically used on override-redirect windows.

PopupMenu

A popup menu that usually appears when the user right clicks on an object. This property is typically used on override-redirect windows.

Tooltip

A tooltip window. Usually used to show additional information when hovering over an object with the cursor. This property is typically used on override-redirect windows.

Notification

The window is a notification. This property is typically used on override-redirect windows.

Combo

This should be used on the windows that are popped up by combo boxes. This property is typically used on override-redirect windows.

Dnd

This indicates the the window is being dragged. This property is typically used on override-redirect windows.

Normal

This is a normal, top-level window.

Implementations

impl WindowType

pub(crate) fn as_atom(&self, xconn: &Arc<XConnection>) -> Atom