Trait winit::platform::unix::EventLoopWindowTargetExtUnix

pub trait EventLoopWindowTargetExtUnix {
    fn is_wayland(&self) -> bool;

    fn is_x11(&self) -> bool;

    fn wayland_display(&self) -> Option<*mut c_void>;
}

Additional methods on EventLoopWindowTarget that are specific to Unix.

Required methods

fn is_wayland(&self) -> bool

True if the EventLoopWindowTarget uses Wayland.

fn is_x11(&self) -> bool

True if the EventLoopWindowTarget uses X11.

fn wayland_display(&self) -> Option<*mut c_void>

Returns a pointer to the wl_display object of wayland that is used by this EventLoopWindowTarget.

Returns None if the EventLoop doesn't use wayland (if it uses xlib for example).

The pointer will become invalid when the winit EventLoop is destroyed.

Loading content...