Trait winit::platform::unix::Theme

pub trait Theme: Send + 'static {
    fn primary_color(&self, window_active: bool) -> [u8; 4];

    fn secondary_color(&self, window_active: bool) -> [u8; 4];

    fn close_button_color(&self, status: ButtonState) -> [u8; 4];

    fn maximize_button_color(&self, status: ButtonState) -> [u8; 4];

    fn minimize_button_color(&self, status: ButtonState) -> [u8; 4];

    fn close_button_icon_color(&self, status: ButtonState) -> [u8; 4] { ... }

    fn maximize_button_icon_color(&self, status: ButtonState) -> [u8; 4] { ... }

    fn minimize_button_icon_color(&self, status: ButtonState) -> [u8; 4] { ... }
}

Required methods

fn primary_color(&self, window_active: bool) -> [u8; 4]

Primary color of the scheme.

fn secondary_color(&self, window_active: bool) -> [u8; 4]

Secondary color of the scheme.

fn close_button_color(&self, status: ButtonState) -> [u8; 4]

Color for the close button.

fn maximize_button_color(&self, status: ButtonState) -> [u8; 4]

Background color for the maximize button.

fn minimize_button_color(&self, status: ButtonState) -> [u8; 4]

Background color for the minimize button.

Loading content...
Provided methods

fn close_button_icon_color(&self, status: ButtonState) -> [u8; 4]

Icon color for the close button, defaults to the secondary color.

fn maximize_button_icon_color(&self, status: ButtonState) -> [u8; 4]

Icon color for the maximize button, defaults to the secondary color.

fn minimize_button_icon_color(&self, status: ButtonState) -> [u8; 4]

Icon color for the minimize button, defaults to the secondary color.

Loading content...