Module winit::platform::wayland

Available on wayland_platform only.

Wayland

Note: Windows don’t appear on Wayland until you draw/present to them.

By default, Winit loads system libraries using dlopen. This can be disabled by disabling the "wayland-dlopen" cargo feature.

Client-side decorations

Winit provides client-side decorations by default, but the behaviour can be controlled with the following feature flags:

wayland-csd-adwaita (default).

wayland-csd-adwaita-crossfont.

wayland-csd-adwaita-notitle.

Re-exports

pub use crate::window::Theme;

Traits

ActiveEventLoopExtWayland

Additional methods on ActiveEventLoop that are specific to Wayland.

EventLoopBuilderExtWayland

Additional methods on EventLoopBuilder that are specific to Wayland.

EventLoopExtWayland

Additional methods on EventLoop that are specific to Wayland.

MonitorHandleExtWayland

Additional methods on MonitorHandle that are specific to Wayland.

WindowAttributesExtWayland

Additional methods on WindowAttributes that are specific to Wayland.

WindowExtWayland

Additional methods on Window that are specific to Wayland.