winit::platform::wayland

Trait WindowExtWayland

pub trait WindowExtWayland {
    // Required method
    fn xdg_toplevel(&self) -> Option<NonNull<c_void>>;
}

Available on wayland_platform only.

Additional methods on Window that are specific to Wayland.

Required Methods

fn xdg_toplevel(&self) -> Option<NonNull<c_void>>

Returns xdg_toplevel of the window or None if the window is X11 window.