winit::platform::web

Trait EventLoopExtWebSys

pub trait EventLoopExtWebSys {
    type UserEvent: 'static;

    // Required methods
    fn spawn_app<A: ApplicationHandler<Self::UserEvent> + 'static>(self, app: A);
    fn spawn<F>(self, event_handler: F)
       where F: 'static + FnMut(Event<Self::UserEvent>, &ActiveEventLoop);
    fn set_poll_strategy(&self, strategy: PollStrategy);
    fn poll_strategy(&self) -> PollStrategy;
    fn set_wait_until_strategy(&self, strategy: WaitUntilStrategy);
    fn wait_until_strategy(&self) -> WaitUntilStrategy;
}

Available on web_platform only.

Additional methods on EventLoop that are specific to the web.

Required Associated Types

type UserEvent: 'static

A type provided by the user that can be passed through Event::UserEvent.

Required Methods

fn spawn_app<A: ApplicationHandler<Self::UserEvent> + 'static>(self, app: A)

Initializes the winit event loop.

Unlike run_app() 1, this returns immediately, and doesn’t throw an exception in order to satisfy its ! return type.

Once the event loop has been destroyed, it’s possible to reinitialize another event loop by calling this function again. This can be useful if you want to recreate the event loop while the WebAssembly module is still loaded. For example, this can be used to recreate the event loop when switching between tabs on a single page application.

run_app() is not available on WASM when the target supports exception-handling. ↩

fn spawn<F>(self, event_handler: F)
where F: 'static + FnMut(Event<Self::UserEvent>, &ActiveEventLoop),

👎Deprecated: use EventLoopExtWebSys::spawn_app

See spawn_app.

fn set_poll_strategy(&self, strategy: PollStrategy)

Sets the strategy for ControlFlow::Poll.

See PollStrategy.

fn poll_strategy(&self) -> PollStrategy

Gets the strategy for ControlFlow::Poll.

See PollStrategy.

fn set_wait_until_strategy(&self, strategy: WaitUntilStrategy)

Sets the strategy for ControlFlow::WaitUntil.

See WaitUntilStrategy.

fn wait_until_strategy(&self) -> WaitUntilStrategy

Gets the strategy for ControlFlow::WaitUntil.

See WaitUntilStrategy.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.