Module winit::platform::x11

Enums

XNotSupported

Error returned if this system doesn’t have XLib or can’t create an X connection.

XWindowType

X window type. Maps directly to _NET_WM_WINDOW_TYPE.

Traits

EventLoopBuilderExtX11

Additional methods on EventLoopBuilder that are specific to X11.

EventLoopWindowTargetExtX11

Additional methods on EventLoopWindowTarget that are specific to X11.

MonitorHandleExtX11

Additional methods on MonitorHandle that are specific to X11.

WindowBuilderExtX11

Additional methods on WindowBuilder that are specific to X11.

WindowExtX11

Additional methods on Window that are specific to X11.

Functions

register_xlib_error_hook

Hook to winit’s xlib error handling callback.

Type Definitions

XlibErrorHook

The first argument in the provided hook will be the pointer to XDisplay and the second one the pointer to XErrorEvent. The returned bool is an indicator whether the error was handled by the callback.