winit::platform::x11

Trait EventLoopBuilderExtX11

pub trait EventLoopBuilderExtX11 {
    // Required methods
    fn with_x11(&mut self) -> &mut Self;
    fn with_any_thread(&mut self, any_thread: bool) -> &mut Self;
}

Available on x11_platform only.

Additional methods on EventLoopBuilder that are specific to X11.

Required Methods

fn with_x11(&mut self) -> &mut Self

Force using X11.

fn with_any_thread(&mut self, any_thread: bool) -> &mut Self

Whether to allow the event loop to be created off of the main thread.

By default, the window is only allowed to be created on the main thread, to make platform compatibility easier.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.