winit::platform_impl::linux::common::xkb

Struct Context

pub struct Context {
    pub core_keyboard_id: i32,
    state: Option<XkbState>,
    keymap: Option<XkbKeymap>,
    compose_state1: Option<XkbComposeState>,
    compose_state2: Option<XkbComposeState>,
    _compose_table: Option<XkbComposeTable>,
    context: XkbContext,
    scratch_buffer: Vec<u8>,
}

Fields
core_keyboard_id: i32state: Option<XkbState>keymap: Option<XkbKeymap>compose_state1: Option<XkbComposeState>compose_state2: Option<XkbComposeState>_compose_table: Option<XkbComposeTable>context: XkbContextscratch_buffer: Vec<u8>
Implementations

impl Context

pub fn new() -> Result<Self, Error>

pub fn from_x11_xkb(xcb: *mut xcb_connection_t) -> Result<Self, Error>

pub fn state_mut(&mut self) -> Option<&mut XkbState>

pub fn keymap_mut(&mut self) -> Option<&mut XkbKeymap>

pub fn set_keymap_from_fd(&mut self, fd: OwnedFd, size: usize)

pub fn set_keymap_from_x11(&mut self, xcb: *mut xcb_connection_t)

pub fn key_context(&mut self) -> Option<KeyContext<'_>>

Key builder context with the user provided xkb state.

pub fn key_context_with_state<'a>( &'a mut self, state: &'a mut XkbState, ) -> Option<KeyContext<'a>>

Key builder context with the user provided xkb state.

Should be used when the original context must not be altered.