Struct winit::platform_impl::linux::wayland::seat::keyboard::KeyboardState

pub struct KeyboardState {
    pub keyboard: WlKeyboard,
    pub loop_handle: LoopHandle<'static, WinitState>,
    pub xkb_context: Context,
    pub repeat_info: RepeatInfo,
    pub repeat_token: Option<RegistrationToken>,
    pub current_repeat: Option<u32>,
}

The state of the keyboard on the current seat.

Fields
keyboard: WlKeyboard

The underlying WlKeyboard.

loop_handle: LoopHandle<'static, WinitState>

Loop handle to handle key repeat.

xkb_context: Context

The state of the keyboard.

repeat_info: RepeatInfo

The information about the repeat rate obtained from the compositor.

repeat_token: Option<RegistrationToken>

The token of the current handle inside the calloop’s event loop.

current_repeat: Option<u32>

The current repeat raw key.

Implementations

impl KeyboardState

pub fn new( keyboard: WlKeyboard, loop_handle: LoopHandle<'static, WinitState>, ) -> Self