Struct winit::platform_impl::linux::wayland::state::WinitState

pub struct WinitState {

    pub registry_state: RegistryState,
    pub output_state: OutputState,
    pub compositor_state: Arc<CompositorState>,
    pub subcompositor_state: Option<Arc<SubcompositorState>>,
    pub seat_state: SeatState,
    pub shm: Shm,
    pub custom_cursor_pool: Arc<Mutex<SlotPool>>,
    pub xdg_shell: XdgShell,
    pub windows: RefCell<AHashMap<WindowId, Arc<Mutex<WindowState>>>>,
    pub window_requests: RefCell<AHashMap<WindowId, Arc<WindowRequests>>>,
    pub window_events_sink: Arc<Mutex<EventSink>>,
    pub window_compositor_updates: Vec<WindowCompositorUpdate>,
    pub seats: AHashMap<ObjectId, WinitSeatState>,
    pub pointer_surfaces: AHashMap<ObjectId, Arc<ThemedPointer<WinitPointerData>>>,
    pub text_input_state: Option<TextInputState>,
    pub monitors: Arc<Mutex<Vec<MonitorHandle>>>,
    pub events_sink: EventSink,
    pub xdg_activation: Option<XdgActivationState>,
    pub relative_pointer: Option<RelativePointerState>,
    pub pointer_constraints: Option<Arc<PointerConstraintsState>>,
    pub viewporter_state: Option<ViewporterState>,
    pub fractional_scaling_manager: Option<FractionalScalingManager>,
    pub kwin_blur_manager: Option<KWinBlurManager>,
    pub loop_handle: LoopHandle<'static, Self>,
    pub dispatched_events: bool,

}

Winit’s Wayland state.

Fields
registry_state: RegistryState

The WlRegistry.

output_state: OutputState

The state of the WlOutput handling.

compositor_state: Arc<CompositorState>

The compositor state which is used to create new windows and regions.

subcompositor_state: Option<Arc<SubcompositorState>>

The state of the subcompositor.

seat_state: SeatState

The seat state responsible for all sorts of input.

shm: Shm

The shm for software buffers, such as cursors.

custom_cursor_pool: Arc<Mutex<SlotPool>>

The pool where custom cursors are allocated.

xdg_shell: XdgShell

The XDG shell that is used for windows.

windows: RefCell<AHashMap<WindowId, Arc<Mutex<WindowState>>>>

The currently present windows.

window_requests: RefCell<AHashMap<WindowId, Arc<WindowRequests>>>

The requests from the Window to EventLoop, such as close operations and redraw requests.

window_events_sink: Arc<Mutex<EventSink>>

The events that were generated directly from the window.

window_compositor_updates: Vec<WindowCompositorUpdate>

The update for the windows coming from the compositor.

seats: AHashMap<ObjectId, WinitSeatState>

Currently handled seats.

pointer_surfaces: AHashMap<ObjectId, Arc<ThemedPointer<WinitPointerData>>>

Currently present cursor surfaces.

text_input_state: Option<TextInputState>

The state of the text input on the client.

monitors: Arc<Mutex<Vec<MonitorHandle>>>

Observed monitors.

events_sink: EventSink

Sink to accumulate window events from the compositor, which is latter dispatched in event loop run.

xdg_activation: Option<XdgActivationState>

Xdg activation.

relative_pointer: Option<RelativePointerState>

Relative pointer.

pointer_constraints: Option<Arc<PointerConstraintsState>>

Pointer constraints to handle pointer locking and confining.

viewporter_state: Option<ViewporterState>

Viewporter state on the given window.

fractional_scaling_manager: Option<FractionalScalingManager>

Fractional scaling manager.

kwin_blur_manager: Option<KWinBlurManager>

KWin blur manager.

loop_handle: LoopHandle<'static, Self>

Loop handle to re-register event sources, such as keyboard repeat.

dispatched_events: bool

Whether we have dispatched events to the user thus we want to send AboutToWait and normally wakeup the user.

Implementations

impl WinitState

fn on_keyboard_destroy(&mut self, seat: &ObjectId)

impl WinitState

pub fn new( globals: &GlobalList, queue_handle: &QueueHandle<Self>, loop_handle: LoopHandle<'static, WinitState>, ) -> Result<Self, OsError>

pub fn scale_factor_changed( &mut self, surface: &WlSurface, scale_factor: f64, is_legacy: bool, )

pub fn queue_close( updates: &mut Vec<WindowCompositorUpdate>, window_id: WindowId, )