winit::platform_impl::linux::wayland

Module window

The Wayland window.

Re-exports

pub use state::WindowState;

Modules

state 🔒

The state of the window, which is shared with the event-loop.

Structs

Window

The Wayland window.

WindowRequests

The request from the window to the event loop.