winit::platform_impl::linux::x11::ime

Module context

Structs

ImeContext

ImeContextClientData 🔒

PreeditCallbacks 🔒

Struct to simplify callback creation and latter passing into Xlib XIM.

Enums

ImeContextCreationError

IME creation error.

Functions

calc_byte_position 🔒

create_xim_callback 🔒

Wrapper for creating XIM callbacks.

preedit_caret_callback 🔒

Handling of cursor movements in preedit text.

preedit_done_callback 🔒

Done callback is used when the preedit should be hidden.

preedit_draw_callback 🔒

Preedit text information to be drawn inline by the client.

preedit_start_callback 🔒

The server started preedit.

Type Aliases

XIMProcNonnull 🔒

The callback used by XIM preedit functions.