winit::platform_impl::linux::x11

Module window

Macros

leap 🔒

Structs

SharedState

UnownedWindow

Enums

Visibility

Functions

cast_dimension_to_hint 🔒

Cast a dimension value into a hinted dimension for WmSizeHints, clamping if too large.

cast_physical_size_to_hint 🔒

Use the above strategy to cast a physical size into a hinted size.

cast_size_to_hint 🔒

Use the above strategy to cast a size into a hinted size.