Struct winit::platform_impl::platform::wayland::seat::pointer::WinitPointer

pub struct WinitPointer {
    pointer: ThemedPointer,
    pointer_constraints: Option<Attached<ZwpPointerConstraintsV1>>,
    confined_pointer: Weak<RefCell<Option<ZwpConfinedPointerV1>>>,
    latest_serial: Rc<Cell<u32>>,
}

A proxy to Wayland pointer, which serves requests from a WindowHandle.

Fields
pointer: ThemedPointerpointer_constraints: Option<Attached<ZwpPointerConstraintsV1>>

Create confined pointers.

confined_pointer: Weak<RefCell<Option<ZwpConfinedPointerV1>>>

Cursor to handle confine requests.

latest_serial: Rc<Cell<u32>>

Latest observed serial in pointer events.

Implementations

impl WinitPointer

pub fn set_cursor(&self, cursor_icon: Option<CursorIcon>)

Set the cursor icon.

Providing None will hide the cursor.

pub fn confine(&self, surface: &WlSurface)

Confine the pointer to a surface.

pub fn unconfine(&self)

Tries to unconfine the pointer if the current pointer is confined.