Struct winit::platform_impl::platform::x11::monitor::MonitorHandle

pub struct MonitorHandle {
    pub(crate) id: RRCrtc,
    pub(crate) name: String,
    dimensions: (u32, u32),
    position: (i32, i32),
    primary: bool,
    refresh_rate_millihertz: Option<u32>,
    pub(crate) scale_factor: f64,
    pub(crate) rect: AaRect,
    video_modes: Vec<VideoMode>,
}

Fields
id: RRCrtc

The actual id

name: String

The name of the monitor

dimensions: (u32, u32)

The size of the monitor

position: (i32, i32)

The position of the monitor in the X screen

primary: bool

If the monitor is the primary one

refresh_rate_millihertz: Option<u32>

The refresh rate used by monitor.

scale_factor: f64

The DPI scale factor

rect: AaRect

Used to determine which windows are on this monitor

video_modes: Vec<VideoMode>

Supported video modes on this monitor

Implementations

impl MonitorHandle

fn new( xconn: &XConnection, resources: *mut XRRScreenResources, id: RRCrtc, crtc: *mut XRRCrtcInfo, primary: bool ) -> Option<Self>

pub fn dummy() -> Self

pub(crate) fn is_dummy(&self) -> bool

pub fn name(&self) -> Option<String>

pub fn native_identifier(&self) -> u32

pub fn size(&self) -> PhysicalSize<u32>

pub fn position(&self) -> PhysicalPosition<i32>

pub fn refresh_rate_millihertz(&self) -> Option<u32>

pub fn scale_factor(&self) -> f64

pub fn video_modes(&self) -> impl Iterator<Item = PlatformVideoMode>