Struct winit::KeyboardInput

pub struct KeyboardInput {
    pub scancode: ScanCode,
    pub state: ElementState,
    pub virtual_keycode: Option<VirtualKeyCode>,
    pub modifiers: ModifiersState,
}

Describes a keyboard input event.

Fields
scancode: ScanCode

Identifies the physical key pressed

This should not change if the user adjusts the host's keyboard map. Use when the physical location of the key is more important than the key's host GUI semantics, such as for movement controls in a first-person game.

state: ElementState virtual_keycode: Option<VirtualKeyCode>

Identifies the semantic meaning of the key

Use when the semantics of the key are more important than the physical location of the key, such as when implementing appropriate behavior for "page up."

modifiers: ModifiersState

Modifier keys active at the time of this input.

This is tracked internally to avoid tracking errors arising from modifier key state changes when events from this device are not being delivered to the application, e.g. due to keyboard focus being elsewhere.