Struct winit::MonitorId

pub struct MonitorId { /* fields omitted */ }

Identifier for a monitor.

Methods

impl MonitorId

pub fn get_name(&self) -> Option<String>

Returns a human-readable name of the monitor.

Returns None if the monitor doesn't exist anymore.

pub fn get_dimensions(&self) -> (u32, u32)

Returns the number of pixels currently displayed on the monitor.

pub fn get_position(&self) -> (i32, i32)

Returns the top-left corner position of the monitor relative to the larger full screen area.

pub fn get_hidpi_factor(&self) -> f32

Returns the ratio between the monitor's physical pixels and logical pixels.