Module winit::window

The Window struct and associated types.

Structs

Icon

An icon used for the window titlebar, taskbar, etc.

Window

Represents a window.

WindowAttributes

Attributes to use when creating a window.

WindowBuilder

Object that allows building windows.

WindowButtons

WindowId

Identifier of a window. Unique for each window.

Enums

BadIcon

An error produced when using Icon::from_rgba with invalid arguments.

CursorGrabMode

The behavior of cursor grabbing.

CursorIcon

Describes the appearance of the mouse cursor.

Fullscreen

Fullscreen modes.

ImePurpose

Generic IME purposes for use in Window::set_ime_purpose.

ResizeDirection

Defines the orientation that a window resize will be performed.

Theme

The theme variant to use.

UserAttentionType

Platform-specific

WindowLevel

A window level groups windows with respect to their z-position.