winit::window

Struct CustomCursor

pub struct CustomCursor { /* private fields */ }

Use a custom image as a cursor (mouse pointer).

Is guaranteed to be cheap to clone.

Platform-specific

Web: Some browsers have limits on cursor sizes usually at 128x128.

Example

use winit::window::CustomCursor;

let w = 10;
let h = 10;
let rgba = vec![255; (w * h * 4) as usize];

#[cfg(not(target_family = "wasm"))]
let source = CustomCursor::from_rgba(rgba, w, h, w / 2, h / 2).unwrap();

#[cfg(target_family = "wasm")]
let source = {
    use winit::platform::web::CustomCursorExtWebSys;
    CustomCursor::from_url(String::from("http://localhost:3000/cursor.png"), 0, 0)
};

let custom_cursor = event_loop.create_custom_cursor(source);

window.set_cursor(custom_cursor.clone());

Implementations

impl CustomCursor

pub fn from_rgba( rgba: impl Into<Vec<u8>>, width: u16, height: u16, hotspot_x: u16, hotspot_y: u16, ) -> Result<CustomCursorSource, BadImage>

Creates a new cursor from an rgba buffer.

The alpha channel is assumed to be not premultiplied.