Struct winit::window::WindowAttributes

pub struct WindowAttributes {
    pub inner_size: Option<Size>,
    pub min_inner_size: Option<Size>,
    pub max_inner_size: Option<Size>,
    pub resizable: bool,
    pub fullscreen: Option<Fullscreen>,
    pub title: String,
    pub maximized: bool,
    pub visible: bool,
    pub transparent: bool,
    pub decorations: bool,
    pub always_on_top: bool,
    pub window_icon: Option<Icon>,
}

Attributes to use when creating a window.

Fields
inner_size: Option<Size>

The dimensions of the window. If this is None, some platform-specific dimensions will be used.

The default is None.

min_inner_size: Option<Size>

The minimum dimensions a window can be, If this is None, the window will have no minimum dimensions (aside from reserved).

The default is None.

max_inner_size: Option<Size>

The maximum dimensions a window can be, If this is None, the maximum will have no maximum or will be set to the primary monitor's dimensions by the platform.

The default is None.

resizable: bool

Whether the window is resizable or not.

The default is true.

fullscreen: Option<Fullscreen>

Whether the window should be set as fullscreen upon creation.

The default is None.

title: String

The title of the window in the title bar.

The default is "winit window".

maximized: bool

Whether the window should be maximized upon creation.

The default is false.

visible: bool

Whether the window should be immediately visible upon creation.

The default is true.

transparent: bool

Whether the the window should be transparent. If this is true, writing colors with alpha values different than 1.0 will produce a transparent window.

The default is false.

decorations: bool

Whether the window should have borders and bars.

The default is true.

always_on_top: bool

Whether the window should always be on top of other windows.

The default is false.

window_icon: Option<Icon>

The window icon.

The default is None.