Struct winit::window::WindowBuilder

pub struct WindowBuilder { /* private fields */ }

Object that allows building windows.

Implementations

impl WindowBuilder

pub fn new() -> Self

Initializes a new builder with default values.

impl WindowBuilder

pub fn window_attributes(&self) -> &WindowAttributes

Get the current window attributes.

pub fn with_inner_size<S: Into<Size>>(self, size: S) -> Self

Requests the window to be of specific dimensions.

If this is not set, some platform-specific dimensions will be used.

See Window::request_inner_size for details.

pub fn with_min_inner_size<S: Into<Size>>(self, min_size: S) -> Self

Sets the minimum dimensions a window can have.

If this is not set, the window will have no minimum dimensions (aside from reserved).

See Window::set_min_inner_size for details.

pub fn with_max_inner_size<S: Into<Size>>(self, max_size: S) -> Self

Sets the maximum dimensions a window can have.

If this is not set, the window will have no maximum or will be set to the primary monitor’s dimensions by the platform.

See Window::set_max_inner_size for details.

pub fn with_position<P: Into<Position>>(self, position: P) -> Self

Sets a desired initial position for the window.

If this is not set, some platform-specific position will be chosen.

See Window::set_outer_position for details.

Platform-specific

macOS: The top left corner position of the window content, the window’s “inner” position. The window title bar will be placed above it. The window will be positioned such that it fits on screen, maintaining set inner_size if any. If you need to precisely position the top left corner of the whole window you have to use Window::set_outer_position after creating the window.

Windows: The top left corner position of the window title bar, the window’s “outer” position. There may be a small gap between this position and the window due to the specifics of the Window Manager.

X11: The top left corner of the window, the window’s “outer” position.

Others: Ignored.

pub fn with_resizable(self, resizable: bool) -> Self

Sets whether the window is resizable or not.

The default is true.

See Window::set_resizable for details.

pub fn with_enabled_buttons(self, buttons: WindowButtons) -> Self

Sets the enabled window buttons.

The default is WindowButtons::all

See Window::set_enabled_buttons for details.

pub fn with_title<T: Into<String>>(self, title: T) -> Self

Sets the initial title of the window in the title bar.

The default is "winit window".

See Window::set_title for details.

pub fn with_fullscreen(self, fullscreen: Option<Fullscreen>) -> Self

Sets whether the window should be put into fullscreen upon creation.

The default is None.

See Window::set_fullscreen for details.

pub fn with_maximized(self, maximized: bool) -> Self

Request that the window is maximized upon creation.

The default is false.

See Window::set_maximized for details.

pub fn with_visible(self, visible: bool) -> Self

Sets whether the window will be initially visible or hidden.

The default is to show the window.

See Window::set_visible for details.

pub fn with_transparent(self, transparent: bool) -> Self

Sets whether the background of the window should be transparent.

If this is true, writing colors with alpha values different than 1.0 will produce a transparent window. On some platforms this is more of a hint for the system and you’d still have the alpha buffer. To control it see Window::set_transparent.

The default is false.

pub fn with_blur(self, blur: bool) -> Self

Sets whether the background of the window should be blurred by the system.

The default is false.

See Window::set_blur for details.

pub fn transparent(&self) -> bool

Get whether the window will support transparency.

pub fn with_decorations(self, decorations: bool) -> Self

Sets whether the window should have a border, a title bar, etc.

The default is true.

See Window::set_decorations for details.

pub fn with_window_level(self, level: WindowLevel) -> Self

Sets the window level.

This is just a hint to the OS, and the system could ignore it.

The default is WindowLevel::Normal.

See WindowLevel for details.

pub fn with_window_icon(self, window_icon: Option<Icon>) -> Self

Sets the window icon.

The default is None.

See Window::set_window_icon for details.

pub fn with_theme(self, theme: Option<Theme>) -> Self

Sets a specific theme for the window.

If None is provided, the window will use the system theme.

The default is None.

Platform-specific

macOS: This is an app-wide setting.

Wayland: This controls only CSD. When using None it’ll try to use dbus to get the system preference. When explicit theme is used, this will avoid dbus all together.

x11: Build window with _GTK_THEME_VARIANT hint set to dark or light.

iOS / Android / Web / x11 / Orbital: Ignored.

pub fn with_resize_increments<S: Into<Size>>(self, resize_increments: S) -> Self

Build window with resize increments hint.

The default is None.

See Window::set_resize_increments for details.

pub fn with_content_protected(self, protected: bool) -> Self

Prevents the window contents from being captured by other apps.

The default is false.

Platform-specific

macOS: if false, NSWindowSharingNone is used but doesn’t completely prevent all apps from reading the window content, for instance, QuickTime.

iOS / Android / Web / x11 / Orbital: Ignored.

pub fn with_active(self, active: bool) -> Self

Whether the window will be initially focused or not.

The window should be assumed as not focused by default following by the WindowEvent::Focused.

Platform-specific:

Android / iOS / X11 / Wayland / Orbital: Unsupported.

pub unsafe fn with_parent_window( self, parent_window: Option<RawWindowHandle> ) -> Self

Available on crate feature rwh_06 only.

Build window with parent window.

The default is None.

Safety

parent_window must be a valid window handle.

Platform-specific

Windows : A child window has the WS_CHILD style and is confined to the client area of its parent window. For more information, see https://docs.microsoft.com/en-us/windows/win32/winmsg/window-features#child-windows

X11: A child window is confined to the client area of its parent window.

Android / iOS / Wayland / Web: Unsupported.

pub fn build<T: 'static>( self, window_target: &EventLoopWindowTarget<T> ) -> Result<Window, OsError>

Builds the window.

Possible causes of error include denied permission, incompatible system, and lack of memory.

Platform-specific

Web: The window is created but not inserted into the web page automatically. Please see the web platform module for more information.