winit::window

Trait PlatformWindowAttributes

pub trait PlatformWindowAttributes:
    AsAny
    + Debug
    + Send
    + Sync {
    // Required method
    fn box_clone(&self) -> Box<dyn PlatformWindowAttributes>;
}

Required Methods

fn box_clone(&self) -> Box<dyn PlatformWindowAttributes>

Implementations

impl dyn PlatformWindowAttributes + '_

pub fn cast_ref<T>(&self) -> Option<&T>
where T: PlatformWindowAttributes,

Downcast to the backend concrete type.

Returns None if the object was not from that backend.

pub fn cast_mut<T>(&mut self) -> Option<&mut T>
where T: PlatformWindowAttributes,

Mutable downcast to the backend concrete type.

Returns None if the object was not from that backend.

pub fn cast<T>( self: Box<dyn PlatformWindowAttributes + '_>, ) -> Result<Box<T>, Box<dyn PlatformWindowAttributes + '_>>
where T: PlatformWindowAttributes,

Owned downcast to the backend concrete type.

Returns Err with self if the object was not from that backend.