Crate winit

Winit allows you to build a window on as many platforms as possible.

Building a window

Before you can build a Window, you first need to build an EventLoop. This is done with the EventLoop::new() function.

use winit::event_loop::EventLoop;
let event_loop = EventLoop::new();

Once this is done there are two ways to create a Window:

Calling Window::new(&event_loop).

Calling let builder = WindowBuilder::new() then builder.build(&event_loop).

The first way is the simplest way and will give you default values for everything.

The second way allows you to customize the way your Window will look and behave by modifying the fields of the WindowBuilder object before you create the Window.

Event handling

Once a Window has been created, it will generate events. For example whenever the user moves the Window, resizes the Window, moves the mouse, etc. an event is generated.

The events generated by a Window can be retreived from the EventLoop the Window was created with.

You do this by calling event_loop.run(...). This function will run forever unless control_flow is set to ControlFlow::Exit, at which point Event::LoopDestroyed is emitted and the entire program terminates.

use winit::event_loop::ControlFlow;
use winit::event::{Event, WindowEvent};

event_loop.run(move |event, _, control_flow| {
    match event {
        Event::WindowEvent { event: WindowEvent::CloseRequested, .. } => {
            println!("The close button was pressed; stopping");
            *control_flow = ControlFlow::Exit
        },
        _ => *control_flow = ControlFlow::Wait,
    }
});

If you use multiple Windows, Event::WindowEvent has a member named window_id. You can compare it with the value returned by the id() method of Window in order to know which Window has received the event.

Drawing on the window

Winit doesn't provide any function that allows drawing on a Window. However it allows you to retrieve the raw handle of the window (see the platform module), which in turn allows you to create an OpenGL/Vulkan/DirectX/Metal/etc. context that will draw on the Window.

Modules

dpi
DPI is important, so read the docs for this module if you don't want to be confused.

event
The Event enum and assorted supporting types.

event_loop
The EventLoop struct and assorted supporting types, including ControlFlow.

monitor
Types useful for interacting with a user's monitors.

platform
Contains traits with platform-specific methods in them.

window
The Window struct and associated types.