Crate winit

The purpose of this library is to provide an OpenGL context on as many platforms as possible.

Building a window

There are two ways to create a window:

Calling Window::new().

Calling let builder = WindowBuilder::new() then builder.build().

The first way is the simpliest way and will give you default values.

The second way allows you to customize the way your window and GL context will look and behave.

Features

This crate has two Cargo features: window and headless.

window allows you to create regular windows and enables the WindowBuilder object.

headless allows you to do headless rendering, and enables the HeadlessRendererBuilder object.

By default only window is enabled.

Modules

os
Contains traits with platform-specific methods in them.

Structs

AvailableMonitorsIter
An iterator for the list of available monitors.

MonitorId
Identifier for a monitor.

PollEventsIterator
An iterator for the poll_events function.

Touch
Represents touch event

WaitEventsIterator
An iterator for the wait_events function.

Window
Represents an OpenGL context and the Window or environment around it.

WindowAttributes
Attributes to use when creating a window.

WindowBuilder
Object that allows you to build windows.

WindowProxy
Represents a thread safe subset of operations that can be called on a window. This structure can be safely cloned and sent between threads.

Enums

CreationError
Error that can happen while creating a window or a headless renderer.

CursorState
Describes how glutin handles the cursor.

ElementState

Event

MouseButton

MouseCursor

MouseScrollDelta

NativeMonitorId
Native platform identifier for a monitor. Different platforms use fundamentally different types to represent a monitor ID.

TouchPhase

VirtualKeyCode

Functions

get_available_monitors
Returns the list of all available monitors.

get_primary_monitor
Returns the primary monitor of the system.

Type Definitions

ScanCode