winit::changelog

Module v0_10

Available nowhere.

0.10.1

Yanked

0.10.0

Add support for Touch for emscripten backend.

Added support for DroppedFile, HoveredFile, and HoveredFileCancelled to X11 backend.

Breaking: unix::WindowExt no longer returns pointers for things that aren’t actually pointers; get_xlib_window now returns Option<std::os::raw::c_ulong> and get_xlib_screen_id returns Option<std::os::raw::c_int>. Additionally, methods that previously returned libc::c_void have been changed to return std::os::raw::c_void, which are not interchangeable types, so users wanting the former will need to explicitly cast.

Added set_decorations method to Window to allow decorations to be toggled after the window is built. Presently only implemented on X11.

Raised the minimum supported version of Rust to 1.20 on MacOS due to usage of associated constants in new versions of cocoa and core-graphics.

Added modifiers field to MouseInput, MouseWheel, and CursorMoved events to track the modifiers state (ModifiersState).

Fixed the emscripten backend to return the size of the canvas instead of the size of the window.