winit::changelog

Module v0_11

Available nowhere.

0.11.3

Added set_min_dimensions and set_max_dimensions methods to Window, and implemented on Windows, X11, Wayland, and OSX.

On X11, dropping a Window actually closes it now, and clicking the window’s × button (or otherwise having the WM signal to close it) will result in the window closing.

Added WindowBuilderExt methods for macos: with_titlebar_transparent, with_title_hidden, with_titlebar_buttons_hidden, with_fullsize_content_view.

Mapped X11 numpad keycodes (arrows, Home, End, PageUp, PageDown, Insert and Delete) to corresponding virtual keycodes

0.11.2

Impl Hash, PartialEq, and Eq for events::ModifiersState.

Implement MonitorId::get_hidpi_factor for MacOS.

Added method os::macos::MonitorIdExt::get_nsscreen() -> *mut c_void that gets a NSScreen object matching the monitor ID.

Send Awakened event on Android when event loop is woken up.

0.11.1

Fixed windows not receiving mouse events when click-dragging the mouse outside the client area of a window, on Windows platforms.

Added method os::android::EventsLoopExt:set_suspend_callback(Option<Box<Fn(bool) -> ()>>) that allows glutin to register a callback when a suspend event happens

0.11.0

Implement MonitorId::get_dimensions for Android.

Added method os::macos::WindowBuilderExt::with_movable_by_window_background(bool) that allows to move a window without a titlebar - with_decorations(false)

Implement Window::set_fullscreen, Window::set_maximized and Window::set_decorations for Wayland.

Added Caret as VirtualKeyCode and support OSX ^-Key with german input.