winit::changelog

Module v0_13

Available nowhere.

0.13.1

Ensure necessary x11-dl version is used.

0.13.0

Implement WindowBuilder::with_maximized, Window::set_fullscreen, Window::set_maximized and Window::set_decorations for MacOS.

Implement WindowBuilder::with_maximized, Window::set_fullscreen, Window::set_maximized and Window::set_decorations for Windows.

On Windows, WindowBuilder::with_fullscreen no longer changing monitor display resolution.

Overhauled X11 window geometry calculations. get_position and set_position are more universally accurate across different window managers, and get_outer_size actually works now.

Fixed SIGSEGV/SIGILL crashes on macOS caused by stabilization of the ! (never) type.

Implement WindowEvent::HiDPIFactorChanged for macOS

On X11, input methods now work completely out of the box, no longer requiring application developers to manually call setlocale. Additionally, when input methods are started, stopped, or restarted on the server end, it’s correctly handled.

Implemented Refresh event on Windows.

Properly calculate the minimum and maximum window size on Windows, including window decorations.

Map more MouseCursor variants to cursor icons on Windows.

Corrected get_position on macOS to return outer frame position, not content area position.

Corrected set_position on macOS to set outer frame position, not content area position.

Added get_inner_position method to Window, which gets the position of the window’s client area. This is implemented on all applicable platforms (all desktop platforms other than Wayland, where this isn’t possible).

Breaking: the Closed event has been replaced by CloseRequested and Destroyed. To migrate, you typically just need to replace all usages of Closed with CloseRequested; see example programs for more info. The exception is iOS, where Closed must be replaced by Destroyed.