winit::changelog

Module v0_17

Available nowhere.

0.17.2

On macOS, fix <C-Tab> so applications receive the event.

On macOS, fix <Cmd-{key}> so applications receive the event.

On Wayland, key press events will now be repeated.

0.17.1

On X11, prevent a compilation failure in release mode for versions of Rust greater than or equal to 1.30.

Fixed deadlock that broke fullscreen mode on Windows.

0.17.0

Cocoa and core-graphics updates.

Fixed thread-safety issues in several Window functions on Windows.

On MacOS, the key state for modifiers key events is now properly set.

On iOS, the view is now set correctly. This makes it possible to render things (instead of being stuck on a black screen), and touch events work again.

Added NetBSD support.

Breaking: On iOS, UIView is now the default root view. WindowBuilderExt::with_root_view_class can be used to set the root view objective-c class to GLKView (OpenGLES) or MTKView (Metal/MoltenVK).

On iOS, the UIApplication is not started until Window::new is called.

Fixed thread unsafety with cursor hiding on macOS.

On iOS, fixed the size of the JmpBuf type used for setjmp/longjmp calls. Previously this was a buffer overflow on most architectures.

On Windows, use cached window DPI instead of repeatedly querying the system. This fixes sporadic crashes on Windows 7.