winit::changelog

Module v0_8

Available nowhere.

0.8.3

Fixed issue of calls to set_inner_size blocking on Windows.

Mapped ISO_Left_Tab to VirtualKeyCode::Tab to make the key work with modifiers

Fixed the X11 backed on 32bit targets

0.8.2

Uniformize keyboard scancode values across Wayland and X11 (#297).

Internal rework of the wayland event loop

Added method os::linux::WindowExt::is_ready

0.8.1

Added various methods to os::linux::EventsLoopExt, plus some hidden items necessary to make glutin work.

0.8.0

Added Window::set_maximized, WindowAttributes::maximized and WindowBuilder::with_maximized.

Added Window::set_fullscreen.

Changed with_fullscreen to take a Option<MonitorId> instead of a MonitorId.

Removed MonitorId::get_native_identifier() in favor of platform-specific traits in the os module.

Changed get_available_monitors() and get_primary_monitor() to be methods of EventsLoop instead of stand-alone methods.

Changed EventsLoop to be tied to a specific X11 or Wayland connection.

Added a os::linux::EventsLoopExt trait that makes it possible to configure the connection.

Fixed the emscripten code, which now compiles.

Changed the X11 fullscreen code to use xrandr instead of xxf86vm.

Fixed the Wayland backend to produce Refresh event after window creation.

Changed the Suspended event to be outside of WindowEvent.

Fixed the X11 backend sometimes reporting the wrong virtual key (#273).