winit::changelog

Module v0_9

Available nowhere.

0.9.0

Added event WindowEvent::HiDPIFactorChanged.

Added method MonitorId::get_hidpi_factor.

Deprecated get_inner_size_pixels and get_inner_size_points methods of Window in favor of get_inner_size.

Breaking: EventsLoop is !Send and !Sync because of platform-dependant constraints, but Window, WindowId, DeviceId and MonitorId guaranteed to be Send.

MonitorId::get_position now returns (i32, i32) instead of (u32, u32).

Rewrite of the wayland backend to use wayland-client-0.11

Support for dead keys on wayland for keyboard utf8 input

Monitor enumeration on Windows is now implemented using EnumDisplayMonitors instead of EnumDisplayDevices. This changes the value returned by MonitorId::get_name().

On Windows added MonitorIdExt::hmonitor method

Impl Clone for EventsLoopProxy

EventsLoop::get_primary_monitor() on X11 will fallback to any available monitor if no primary is found

Support for touch event on wayland

WindowEvents MouseMoved, MouseEntered, and MouseLeft have been renamed to CursorMoved, CursorEntered, and CursorLeft.

New DeviceEvents added, MouseMotion and MouseWheel.

Send CursorMoved event after CursorEntered and Focused events.

Add support for ModifiersState, MouseMove, MouseInput, MouseMotion for emscripten backend.