winit

Module cursor

Structs

CursorImage 🔒

CustomCursor

Use a custom image as a cursor (mouse pointer).

CustomCursorSource

Source for CustomCursor.

NoCustomCursor 🔒

OnlyCursorImage 🔒

Platforms export this directly as PlatformCustomCursor if they don’t implement caching.

OnlyCursorImageSource 🔒

Platforms export this directly as PlatformCustomCursorSource if they need to only work with images.

Enums

BadImage

An error produced when using CustomCursor::from_rgba with invalid arguments.

Cursor

See Window::set_cursor() for more details.

Constants

MAX_CURSOR_SIZE

The maximum width and height for a cursor when using CustomCursor::from_rgba.

PIXEL_SIZE 🔒