Enum winit::WindowEvent

pub enum WindowEvent {
    Resized(LogicalSize),
    Moved(LogicalPosition),
    CloseRequested,
    Destroyed,
    DroppedFile(PathBuf),
    HoveredFile(PathBuf),
    HoveredFileCancelled,
    ReceivedCharacter(char),
    Focused(bool),
    KeyboardInput {
        device_id: DeviceId,
        input: KeyboardInput,
    },
    CursorMoved {
        device_id: DeviceId,
        position: LogicalPosition,
        modifiers: ModifiersState,
    },
    CursorEntered {
        device_id: DeviceId,
    },
    CursorLeft {
        device_id: DeviceId,
    },
    MouseWheel {
        device_id: DeviceId,
        delta: MouseScrollDelta,
        phase: TouchPhase,
        modifiers: ModifiersState,
    },
    MouseInput {
        device_id: DeviceId,
        state: ElementState,
        button: MouseButton,
        modifiers: ModifiersState,
    },
    TouchpadPressure {
        device_id: DeviceId,
        pressure: f32,
        stage: i64,
    },
    AxisMotion {
        device_id: DeviceId,
        axis: AxisId,
        value: f64,
    },
    Refresh,
    Touch(Touch),
    HiDpiFactorChanged(f64),
}

Describes an event from a Window.

Variants
Resized(LogicalSize)

The size of the window has changed. Contains the client area's new dimensions.

Moved(LogicalPosition)

The position of the window has changed. Contains the window's new position.

CloseRequested

The window has been requested to close.

Destroyed

The window has been destroyed.

DroppedFile(PathBuf)

A file has been dropped into the window.

HoveredFile(PathBuf)

A file is being hovered over the window.

HoveredFileCancelled

A file was hovered, but has exited the window.

ReceivedCharacter(char)

The window received a unicode character.

Focused(bool)

The window gained or lost focus.

The parameter is true if the window has gained focus, and false if it has lost focus.

KeyboardInput

An event from the keyboard has been received.

Fields of KeyboardInput

device_id: DeviceIdinput: KeyboardInput
CursorMoved

The cursor has moved on the window.

Fields of CursorMoved

device_id: DeviceIdposition: LogicalPosition

(x,y) coords in pixels relative to the top-left corner of the window. Because the range of this data is limited by the display area and it may have been transformed by the OS to implement effects such as cursor acceleration, it should not be used to implement non-cursor-like interactions such as 3D camera control.

modifiers: ModifiersState
CursorEntered

The cursor has entered the window.

Fields of CursorEntered

device_id: DeviceId
CursorLeft

The cursor has left the window.

Fields of CursorLeft

device_id: DeviceId
MouseWheel

A mouse wheel movement or touchpad scroll occurred.

Fields of MouseWheel

device_id: DeviceIddelta: MouseScrollDeltaphase: TouchPhasemodifiers: ModifiersState
MouseInput

An mouse button press has been received.

Fields of MouseInput

device_id: DeviceIdstate: ElementStatebutton: MouseButtonmodifiers: ModifiersState
TouchpadPressure

Touchpad pressure event.

At the moment, only supported on Apple forcetouch-capable macbooks. The parameters are: pressure level (value between 0 and 1 representing how hard the touchpad is being pressed) and stage (integer representing the click level).

Fields of TouchpadPressure

device_id: DeviceIdpressure: f32stage: i64
AxisMotion

Motion on some analog axis. May report data redundant to other, more specific events.

Fields of AxisMotion

device_id: DeviceIdaxis: AxisIdvalue: f64
Refresh

The window needs to be redrawn.

Touch(Touch)

Touch event has been received

HiDpiFactorChanged(f64)

The DPI factor of the window has changed.

The following user actions can cause DPI changes:

Changing the display's resolution.

Changing the display's DPI factor (e.g. in Control Panel on Windows).

Moving the window to a display with a different DPI factor.

For more information about DPI in general, see the dpi module.