Module winit::event

The Event enum and assorted supporting types.

These are sent to the closure given to EventLoop::run(...), where they get processed and used to modify the program state. For more details, see the root-level documentation.

Modules

keyboard
Keyboard event types.

Structs

DeviceId
Identifier of an input device.

Touch
Represents touch event

Enums

DeviceEvent
Represents raw hardware events that are not associated with any particular window.

ElementState
Describes the input state of a key.

Event
Describes a generic event.

MouseButton
Describes a button of a mouse controller.

MouseScrollDelta
Describes a difference in the mouse scroll wheel state.

StartCause
Describes the reason the event loop is resuming.

TouchPhase
Describes touch-screen input state.

WindowEvent
Describes an event from a Window.

Type Definitions

AxisId
Identifier for a specific analog axis on some device.

ButtonId
Identifier for a specific button on some device.

ScanCode
Hardware-dependent keyboard scan code.