winit::event

Enum WindowEvent

pub enum WindowEvent {

    ActivationTokenDone {
        serial: AsyncRequestSerial,
        token: ActivationToken,
    },
    Resized(PhysicalSize<u32>),
    Moved(PhysicalPosition<i32>),
    CloseRequested,
    Destroyed,
    DroppedFile(PathBuf),
    HoveredFile(PathBuf),
    HoveredFileCancelled,
    Focused(bool),
    KeyboardInput {
        device_id: DeviceId,
        event: KeyEvent,
        is_synthetic: bool,
    },
    ModifiersChanged(Modifiers),
    Ime(Ime),
    CursorMoved {
        device_id: DeviceId,
        position: PhysicalPosition<f64>,
    },
    CursorEntered {
        device_id: DeviceId,
    },
    CursorLeft {
        device_id: DeviceId,
    },
    MouseWheel {
        device_id: DeviceId,
        delta: MouseScrollDelta,
        phase: TouchPhase,
    },
    MouseInput {
        device_id: DeviceId,
        state: ElementState,
        button: MouseButton,
    },
    PinchGesture {
        device_id: DeviceId,
        delta: f64,
        phase: TouchPhase,
    },
    PanGesture {
        device_id: DeviceId,
        delta: PhysicalPosition<f32>,
        phase: TouchPhase,
    },
    DoubleTapGesture {
        device_id: DeviceId,
    },
    RotationGesture {
        device_id: DeviceId,
        delta: f32,
        phase: TouchPhase,
    },
    TouchpadPressure {
        device_id: DeviceId,
        pressure: f32,
        stage: i64,
    },
    AxisMotion {
        device_id: DeviceId,
        axis: AxisId,
        value: f64,
    },
    Touch(Touch),
    ScaleFactorChanged {
        scale_factor: f64,
        inner_size_writer: InnerSizeWriter,
    },
    ThemeChanged(Theme),
    Occluded(bool),
    RedrawRequested,

}

Describes an event from a Window.

Variants

ActivationTokenDone

The activation token was delivered back and now could be used. Delivered in response to request_activation_token.

Fields

serial: AsyncRequestSerial

token: ActivationToken

Resized(PhysicalSize<u32>)

The size of the window has changed. Contains the client area’s new dimensions.

Moved(PhysicalPosition<i32>)

The position of the window has changed. Contains the window’s new position.

Platform-specific

iOS / Android / Web / Wayland: Unsupported.

CloseRequested

The window has been requested to close.

Destroyed

The window has been destroyed.

DroppedFile(PathBuf)

A file has been dropped into the window.

When the user drops multiple files at once, this event will be emitted for each file separately.

HoveredFile(PathBuf)

A file is being hovered over the window.

When the user hovers multiple files at once, this event will be emitted for each file separately.

HoveredFileCancelled

A file was hovered, but has exited the window.

There will be a single HoveredFileCancelled event triggered even if multiple files were hovered.

Focused(bool)

The window gained or lost focus.

The parameter is true if the window has gained focus, and false if it has lost focus.

KeyboardInput

An event from the keyboard has been received.

Platform-specific

Windows: The shift key overrides NumLock. In other words, while shift is held down, numpad keys act as if NumLock wasn’t active. When this is used, the OS sends fake key events which are not marked as is_synthetic.

Fields

device_id: DeviceId

event: KeyEvent

is_synthetic: bool

If true, the event was generated synthetically by winit in one of the following circumstances:

Synthetic key press events are generated for all keys pressed when a window gains focus. Likewise, synthetic key release events are generated for all keys pressed when a window goes out of focus. Currently, this is only functional on X11 and Windows

Otherwise, this value is always false.

ModifiersChanged(Modifiers)

The keyboard modifiers have changed.

Ime(Ime)

An event from an input method.

Note: You have to explicitly enable this event using Window::set_ime_allowed.

Platform-specific

iOS / Android / Web / Orbital: Unsupported.

CursorMoved

The cursor has moved on the window.

Platform-specific

Web: Doesn’t take into account CSS border, padding, or transform.

Fields

device_id: DeviceId

position: PhysicalPosition<f64>

(x,y) coords in pixels relative to the top-left corner of the window. Because the range of this data is limited by the display area and it may have been transformed by the OS to implement effects such as cursor acceleration, it should not be used to implement non-cursor-like interactions such as 3D camera control.

CursorEntered

The cursor has entered the window.

Platform-specific

Web: Doesn’t take into account CSS border, padding, or transform.

Fields

device_id: DeviceId

CursorLeft

The cursor has left the window.

Platform-specific

Web: Doesn’t take into account CSS border, padding, or transform.

Fields

device_id: DeviceId

MouseWheel

A mouse wheel movement or touchpad scroll occurred.

Fields

device_id: DeviceId

delta: MouseScrollDelta

phase: TouchPhase

MouseInput

An mouse button press has been received.

Fields

device_id: DeviceId

state: ElementState

button: MouseButton

PinchGesture

Two-finger pinch gesture, often used for magnification.

Platform-specific

Only available on macOS and iOS.

On iOS, not recognized by default. It must be enabled when needed.

Fields

device_id: DeviceId

delta: f64

Positive values indicate magnification (zooming in) and negative values indicate shrinking (zooming out).

This value may be NaN.

phase: TouchPhase

PanGesture

N-finger pan gesture

Platform-specific

Only available on iOS.

On iOS, not recognized by default. It must be enabled when needed.

Fields

device_id: DeviceId

delta: PhysicalPosition<f32>

Change in pixels of pan gesture from last update.

phase: TouchPhase

DoubleTapGesture

Double tap gesture.

On a Mac, smart magnification is triggered by a double tap with two fingers on the trackpad and is commonly used to zoom on a certain object (e.g. a paragraph of a PDF) or (sort of like a toggle) to reset any zoom. The gesture is also supported in Safari, Pages, etc.

The event is general enough that its generating gesture is allowed to vary across platforms. It could also be generated by another device.

Unfortunately, neither Windows nor Wayland support this gesture or any other gesture with the same effect.

Platform-specific

Only available on macOS 10.8 and later, and iOS.

On iOS, not recognized by default. It must be enabled when needed.

Fields

device_id: DeviceId

RotationGesture

Two-finger rotation gesture.

Positive delta values indicate rotation counterclockwise and negative delta values indicate rotation clockwise.

Platform-specific

Only available on macOS and iOS.

On iOS, not recognized by default. It must be enabled when needed.

Fields

device_id: DeviceId

delta: f32

change in rotation in degrees

phase: TouchPhase

TouchpadPressure

Touchpad pressure event.

At the moment, only supported on Apple forcetouch-capable macbooks. The parameters are: pressure level (value between 0 and 1 representing how hard the touchpad is being pressed) and stage (integer representing the click level).

Fields

device_id: DeviceId

pressure: f32

stage: i64

AxisMotion

Motion on some analog axis. May report data redundant to other, more specific events.

Fields

device_id: DeviceId

axis: AxisId

value: f64

Touch(Touch)

Touch event has been received

Platform-specific

Web: Doesn’t take into account CSS border, padding, or transform.

macOS: Unsupported.

ScaleFactorChanged

The window’s scale factor has changed.

The following user actions can cause DPI changes:

Changing the display’s resolution.

Changing the display’s scale factor (e.g. in Control Panel on Windows).

Moving the window to a display with a different scale factor.

To update the window size, use the provided InnerSizeWriter handle. By default, the window is resized to the value suggested by the OS, but it can be changed to any value.

For more information about DPI in general, see the dpi crate.

Fields

scale_factor: f64

inner_size_writer: InnerSizeWriter

Handle to update inner size during scale changes.

See InnerSizeWriter docs for more details.

ThemeChanged(Theme)

The system window theme has changed.

Applications might wish to react to this to change the theme of the content of the window when the system changes the window theme.

This only reports a change if the window theme was not overridden by Window::set_theme.

Platform-specific

iOS / Android / X11 / Wayland / Orbital: Unsupported.

Occluded(bool)

The window has been occluded (completely hidden from view).

This is different to window visibility as it depends on whether the window is closed, minimised, set invisible, or fully occluded by another window.

Platform-specific

iOS

On iOS, the Occluded(false) event is emitted in response to an applicationWillEnterForeground callback which means the application should start preparing its data. The Occluded(true) event is emitted in response to an applicationDidEnterBackground callback which means the application should free resources (according to the iOS application lifecycle).

Others

Web: Doesn’t take into account CSS border, padding, or transform.

Android / Wayland / Windows / Orbital: Unsupported.

RedrawRequested

Emitted when a window should be redrawn.

This gets triggered in two scenarios:

The OS has performed an operation that’s invalidated the window’s contents (such as resizing the window).

The application has explicitly requested a redraw via Window::request_redraw.

Winit will aggregate duplicate redraw requests into a single event, to help avoid duplicating rendering work.