Module winit::event_loop

The EventLoop struct and assorted supporting types, including ControlFlow.

If you want to send custom events to the event loop, use EventLoop::create_proxy to acquire an EventLoopProxy and call its send_event method.

See the root-level documentation for information on how to create and use an event loop to handle events.

Structs

AsyncRequestSerial

A unique identifier of the winit’s async request.

EventLoop

Provides a way to retrieve events from the system and from the windows that were registered to the events loop.

EventLoopBuilder

Object that allows building the event loop.

EventLoopClosed

The error that is returned when an EventLoopProxy attempts to wake up an EventLoop that no longer exists.

EventLoopProxy

Used to send custom events to EventLoop.

EventLoopWindowTarget

Target that associates windows with an EventLoop.

Enums

ControlFlow

Set through EventLoopWindowTarget::set_control_flow().

DeviceEvents

Control when device events are captured.