winit::event_loop

Struct EventLoop

pub struct EventLoop<T: 'static> { /* private fields */ }

Provides a way to retrieve events from the system and from the windows that were registered to the events loop.

An EventLoop can be seen more or less as a โcontextโ. Calling EventLoop::new initializes everything that will be required to create windows. For example on Linux creating an event loop opens a connection to the X or Wayland server.

To wake up an EventLoop from a another thread, see the EventLoopProxy docs.

Note that this cannot be shared across threads (due to platform-dependant logic forbidding it), as such it is neither Send nor Sync. If you need cross-thread access, the Window created from this can be sent to an other thread, and the EventLoopProxy allows you to wake up an EventLoop from another thread.

Implementations

impl EventLoop<()>

pub fn new() -> Result<EventLoop<()>, EventLoopError>

Create the event loop.

This is an alias of EventLoop::builder().build().

pub fn builder() -> EventLoopBuilder<()>

Start building a new event loop.

This returns an EventLoopBuilder, to allow configuring the event loop before creation.

To get the actual event loop, call build on that.

impl<T> EventLoop<T>

pub fn with_user_event() -> EventLoopBuilder<T>

Start building a new event loop, with the given type as the user event type.

pub fn run<F>(self, event_handler: F) -> Result<(), EventLoopError>
where F: FnMut(Event<T>, &ActiveEventLoop),

๐Deprecated: use EventLoop::run_app instead

See run_app.

pub fn run_app<A: ApplicationHandler<T>>( self, app: &mut A, ) -> Result<(), EventLoopError>

Run the application with the event loop on the calling thread.

See the set_control_flow() docs on how to change the event loopโs behavior.

Platform-specific

iOS: Will never return to the caller and so values not passed to this function will not be dropped before the process exits.

Web: Will act as if it never returns to the caller by throwing a Javascript exception (that Rust doesnโt see) that will also mean that the rest of the function is never executed and any values not passed to this function will not be dropped.

Web applications are recommended to use EventLoopExtWebSys::spawn() 1 instead of run_app() to avoid the need for the Javascript exception trick, and to make it clearer that the event loop runs asynchronously (via the browserโs own, internal, event loop) and doesnโt block the current thread of execution like it does on other platforms.

This function wonโt be available with target_feature = "exception-handling".

EventLoopExtWebSys::spawn_app() is only available on Web. โฉ

pub fn create_proxy(&self) -> EventLoopProxy<T>

Creates an EventLoopProxy that can be used to dispatch user events to the main event loop, possibly from another thread.

pub fn owned_display_handle(&self) -> OwnedDisplayHandle

Gets a persistent reference to the underlying platform display.

See the OwnedDisplayHandle type for more information.

pub fn listen_device_events(&self, allowed: DeviceEvents)

Change if or when DeviceEvents are captured.

See ActiveEventLoop::listen_device_events for details.

pub fn set_control_flow(&self, control_flow: ControlFlow)

Sets the ControlFlow.

pub fn create_window( &self, window_attributes: WindowAttributes, ) -> Result<Window, OsError>

๐Deprecated: use ActiveEventLoop::create_window instead

Create a window.

Creating window without event loop running often leads to improper window creation; use ActiveEventLoop::create_window instead.

pub fn create_custom_cursor( &self, custom_cursor: CustomCursorSource, ) -> CustomCursor

Create custom cursor.