Struct winit::event_loop::EventLoop

pub struct EventLoop<T: 'static> { /* fields omitted */ }

Provides a way to retrieve events from the system and from the windows that were registered to the events loop.

An EventLoop can be seen more or less as a "context". Calling EventLoop::new() initializes everything that will be required to create windows. For example on Linux creating an events loop opens a connection to the X or Wayland server.

To wake up an EventLoop from a another thread, see the EventLoopProxy docs.

Note that the EventLoop cannot be shared across threads (due to platform-dependant logic forbidding it), as such it is neither Send nor Sync. If you need cross-thread access, the Window created from this EventLoop can be sent to an other thread, and the EventLoopProxy allows you to wake up an EventLoop from an other thread.

Methods

impl EventLoop<()>

pub fn new() -> EventLoop<()>

Builds a new event loop with a () as the user event type.

impl<T> EventLoop<T>

pub fn new_user_event() -> EventLoop<T>

Builds a new event loop.

Usage will result in display backend initialisation, this can be controlled on linux using an environment variable WINIT_UNIX_BACKEND. Legal values are x11 and wayland. If it is not set, winit will try to connect to a wayland connection, and if it fails will fallback on x11. If this variable is set with any other value, winit will panic.

Important traits for AvailableMonitorsIter
impl Iterator for AvailableMonitorsIter type Item = MonitorHandle;

pub fn get_available_monitors(&self) -> AvailableMonitorsIter

Returns the list of all the monitors available on the system.

pub fn get_primary_monitor(&self) -> MonitorHandle

Returns the primary monitor of the system.

pub fn run<F>(self, event_handler: F) -> ! where
F: 'static + FnMut(Event<T>, &EventLoopWindowTarget<T>, &mut ControlFlow),

Hijacks the calling thread and initializes the winit event loop with the provided closure. Since the closure is 'static, it must be a move closure if it needs to access any data from the calling context.

See the ControlFlow docs for information on how changes to &mut ControlFlow impact the event loop's behavior.

Any values not passed to this function will not be dropped.

pub fn create_proxy(&self) -> EventLoopProxy<T>

Creates an EventLoopProxy that can be used to wake up the EventLoop from another thread.