Module winit::monitor

Types useful for interacting with a user's monitors.

If you want to get basic information about a monitor, you can use the MonitorHandle type. This is retreived from one of the following methods, which return an iterator of MonitorHandle:

EventLoop::available_monitors

Window::available_monitors.

Structs

MonitorHandle
Handle to a monitor.

VideoMode
Describes a fullscreen video mode of a monitor.