winit::platform

Module macos

Available on macos_platform only.

macOS / AppKit

Winit has an OS requirement of macOS 10.11 or higher (same as Rust itself), and is regularly tested on macOS 10.14.

A lot of functionality expects the application to be ready before you start doing anything; this includes creating windows, fetching monitors, drawing, and so on, see issues #2238, #2051 and #2087.

If you encounter problems, you should try doing your initialization inside Event::Resumed.

Enums

ActivationPolicy

Corresponds to NSApplicationActivationPolicy.

OptionAsAlt

Option as alt behavior.

Traits

ActiveEventLoopExtMacOS

Additional methods on ActiveEventLoop that are specific to macOS.

EventLoopBuilderExtMacOS

MonitorHandleExtMacOS

Additional methods on MonitorHandle that are specific to MacOS.

WindowAttributesExtMacOS

Additional methods on WindowAttributes that are specific to MacOS.

WindowExtMacOS

Additional methods on Window that are specific to MacOS.