Trait winit::platform::macos::WindowExtMacOS

pub trait WindowExtMacOS {
    fn ns_window(&self) -> *mut c_void;

    fn ns_view(&self) -> *mut c_void;

    fn simple_fullscreen(&self) -> bool;

    fn set_simple_fullscreen(&self, fullscreen: bool) -> bool;

    fn has_shadow(&self) -> bool;

    fn set_has_shadow(&self, has_shadow: bool);
}

Additional methods on Window that are specific to MacOS.

Required methods

fn ns_window(&self) -> *mut c_void

Returns a pointer to the cocoa NSWindow that is used by this window.

The pointer will become invalid when the Window is destroyed.

fn ns_view(&self) -> *mut c_void

Returns a pointer to the cocoa NSView that is used by this window.

The pointer will become invalid when the Window is destroyed.

fn simple_fullscreen(&self) -> bool

Returns whether or not the window is in simple fullscreen mode.

fn set_simple_fullscreen(&self, fullscreen: bool) -> bool

Toggles a fullscreen mode that doesn't require a new macOS space. Returns a boolean indicating whether the transition was successful (this won't work if the window was already in the native fullscreen).

This is how fullscreen used to work on macOS in versions before Lion. And allows the user to have a fullscreen window without using another space or taking control over the entire monitor.

fn has_shadow(&self) -> bool

Returns whether or not the window has shadow.

fn set_has_shadow(&self, has_shadow: bool)

Sets whether or not the window has shadow.

Loading content...