Trait winit::platform::startup_notify::WindowAttributesExtStartupNotify

pub trait WindowAttributesExtStartupNotify {
    // Required method
    fn with_activation_token(self, token: ActivationToken) -> Self;
}

Available on x11_platform or wayland_platform only.

Required Methods

fn with_activation_token(self, token: ActivationToken) -> Self

Use this ActivationToken during window creation.

Not using such a token upon a window could make your window not gaining focus until the user clicks on the window.

Object Safety

This trait is not object safe.