Module winit::platform::unix

Re-exports

pub use crate::window::Theme;

Enums

XNotSupported

Error returned if this system doesn’t have XLib or can’t create an X connection.

XWindowType

X window type. Maps directly to _NET_WM_WINDOW_TYPE.

Traits

EventLoopBuilderExtUnix

Additional methods on EventLoopBuilder that are specific to Unix.

EventLoopWindowTargetExtUnix

Additional methods on EventLoopWindowTarget that are specific to Unix.

MonitorHandleExtUnix

Additional methods on MonitorHandle that are specific to Linux.

WindowBuilderExtUnix

Additional methods on WindowBuilder that are specific to Unix.

WindowExtUnix

Additional methods on Window that are specific to Unix.

Functions

register_xlib_error_hook

Hook to winit’s xlib error handling callback.

Type Aliases

XlibErrorHook

The first argument in the provided hook will be the pointer to XDisplay and the second one the pointer to XErrorEvent. The returned bool is an indicator whether the error was handled by the callback.