winit::platform::wayland
Trait EventLoopBuilderExtWayland

pub trait EventLoopBuilderExtWayland {
    // Required methods
    fn with_wayland(&mut self) -> &mut Self;
    fn with_any_thread(&mut self, any_thread: bool) -> &mut Self;
}

Available on wayland_platform only.

Additional methods on EventLoopBuilder that are specific to Wayland.

Required Methods

fn with_wayland(&mut self) -> &mut Self

Force using Wayland.

fn with_any_thread(&mut self, any_thread: bool) -> &mut Self

Whether to allow the event loop to be created off of the main thread.

By default, the window is only allowed to be created on the main thread, to make platform compatibility easier.

Object Safety

This trait is not object safe.