winit::platform

Module windows

Available on windows_platform only.

Windows

The supported OS version is Windows 7 or higher, though Windows 10 is tested regularly.

Structs

AnyThread

A wrapper around a Window that ignores thread-specific window handle limitations.

Color

Describes a color used by Windows

Enums

BackdropType

Describes a system-drawn backdrop material of a window.

CornerPreference

Describes how the corners of a window should look like.

Traits

DeviceIdExtWindows

Additional methods on DeviceId that are specific to Windows.

EventLoopBuilderExtWindows

Additional methods on EventLoop that are specific to Windows.

IconExtWindows

Additional methods on Icon that are specific to Windows.

MonitorHandleExtWindows

Additional methods on MonitorHandle that are specific to Windows.

WindowAttributesExtWindows

Additional methods on WindowAttributes that are specific to Windows.

WindowBorrowExtWindows

Additional methods for anything that dereference to Window.

WindowExtWindows

Additional methods on Window that are specific to Windows.

Type Aliases

HMENU

Menu Handle type used by Win32 API

HMONITOR

Monitor Handle type used by Win32 API

HWND

Window Handle type used by Win32 API