winit::platform::windows

Trait IconExtWindows

pub trait IconExtWindows: Sized {
    // Required methods
    fn from_path<P: AsRef<Path>>(
        path: P,
        size: Option<PhysicalSize<u32>>,
    ) -> Result<Self, BadIcon>;
    fn from_resource(
        ordinal: u16,
        size: Option<PhysicalSize<u32>>,
    ) -> Result<Self, BadIcon>;
}

Available on windows_platform only.

Additional methods on Icon that are specific to Windows.

Required Methods

fn from_path<P: AsRef<Path>>( path: P, size: Option<PhysicalSize<u32>>, ) -> Result<Self, BadIcon>

Create an icon from a file path.

Specify size to load a specific icon size from the file, or None to load the default icon size from the file.

In cases where the specified size does not exist in the file, Windows may perform scaling to get an icon of the desired size.

fn from_resource( ordinal: u16, size: Option<PhysicalSize<u32>>, ) -> Result<Self, BadIcon>

Create an icon from a resource embedded in this executable or library.

Specify size to load a specific icon size from the file, or None to load the default icon size from the file.

In cases where the specified size does not exist in the file, Windows may perform scaling to get an icon of the desired size.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.