winit::platform

Module x11

Available on x11_platform only.

X11

Enums

WindowType

X window type. Maps directly to _NET_WM_WINDOW_TYPE.

Traits

ActiveEventLoopExtX11

Additional methods on ActiveEventLoop that are specific to X11.

EventLoopBuilderExtX11

Additional methods on EventLoopBuilder that are specific to X11.

EventLoopExtX11

Additional methods on EventLoop that are specific to X11.

MonitorHandleExtX11

Additional methods on MonitorHandle that are specific to X11.

WindowAttributesExtX11

Additional methods on WindowAttributes that are specific to X11.

WindowExtX11

Additional methods on Window that are specific to X11.

Functions

register_xlib_error_hook

Hook to winit’s xlib error handling callback.

Type Aliases

XVisualID

A unique identifier for an X11 visual.

XWindow

A unique identifier for an X11 window.

XlibErrorHook

The first argument in the provided hook will be the pointer to XDisplay and the second one the pointer to XErrorEvent. The returned bool is an indicator whether the error was handled by the callback.