Struct winit::platform_impl::linux::wayland::event_loop::EventLoop

pub struct EventLoop<T: 'static> {
    loop_running: bool,
    buffer_sink: EventSink,
    compositor_updates: Vec<WindowCompositorUpdate>,
    window_ids: Vec<WindowId>,
    user_events_sender: Sender<T>,
    pending_user_events: Rc<RefCell<Vec<T>>>,
    wayland_dispatcher: Dispatcher<'static, WaylandSource<WinitState>, WinitState>,
    connection: Connection,
    window_target: ActiveEventLoop,
    event_loop: EventLoop<'static, WinitState>,
}

The Wayland event loop.

Fields
loop_running: bool

Has run or run_on_demand been called or a call to pump_events that starts the loop

buffer_sink: EventSinkcompositor_updates: Vec<WindowCompositorUpdate>window_ids: Vec<WindowId>user_events_sender: Sender<T>

Sender of user events.

pending_user_events: Rc<RefCell<Vec<T>>>

Pending events from the user.

wayland_dispatcher: Dispatcher<'static, WaylandSource<WinitState>, WinitState>

The Wayland dispatcher to has raw access to the queue when needed, such as when creating a new window.

connection: Connection

Connection to the wayland server.

window_target: ActiveEventLoop

Event loop window target.

event_loop: EventLoop<'static, WinitState>

Calloop’s event loop.

Implementations

impl<T: 'static> EventLoop<T>

pub fn new() -> Result<EventLoop<T>, EventLoopError>

pub fn run_on_demand<F>( &mut self, event_handler: F, ) -> Result<(), EventLoopError>
where F: FnMut(Event<T>, &RootActiveEventLoop),

pub fn pump_events<F>( &mut self, timeout: Option<Duration>, callback: F, ) -> PumpStatus
where F: FnMut(Event<T>, &RootActiveEventLoop),

pub fn poll_events_with_timeout<F>( &mut self, timeout: Option<Duration>, callback: F, )
where F: FnMut(Event<T>, &RootActiveEventLoop),

fn single_iteration<F>(&mut self, callback: &mut F, cause: StartCause)
where F: FnMut(Event<T>, &RootActiveEventLoop),

pub fn create_proxy(&self) -> EventLoopProxy<T>

pub fn window_target(&self) -> &RootActiveEventLoop

fn with_state<'a, U: 'a, F: FnOnce(&'a mut WinitState) -> U>( &'a mut self, callback: F, ) -> U

fn loop_dispatch<D: Into<Option<Duration>>>( &mut self, timeout: D, ) -> IOResult<()>

fn roundtrip(&mut self) -> Result<usize, RootOsError>

fn control_flow(&self) -> ControlFlow

fn exiting(&self) -> bool

fn set_exit_code(&self, code: i32)

fn exit_code(&self) -> Option<i32>