Module winit::platform_impl::platform::x11

Re-exports

pub use self::monitor::MonitorHandle;

pub use self::monitor::VideoMode;

pub use self::window::UnownedWindow;

pub use self::xdisplay::XConnection;

pub use self::xdisplay::XError;

pub use self::xdisplay::XNotSupported;

Modules

dnd

event_processor

events

ffi

ime

monitor

util

window

xdisplay

Structs

Device

DeviceId

DeviceInfo

EventLoop

EventLoopProxy

EventLoopWindowTarget

GenericEventCookie

XEvents of type GenericEvent store their actual data in an XGenericEventCookie data structure. This is a wrapper to extract the cookie from a GenericEvent XEvent and release the cookie data once it has been processed

ScrollAxis

Window

WindowId

XExtension

Enums

ScrollOrientation

Constants

REDRAW_TOKEN

USER_TOKEN

X_TOKEN

Functions

get_xtarget

mkdid

mkwid