winit

Module window

The Window struct and associated types.

Structs

ActivationToken

An opaque token used to activate the Window.

CursorIconParseError

An error which could be returned when parsing CursorIcon.

CustomCursor

Use a custom image as a cursor (mouse pointer).

CustomCursorSource

Source for CustomCursor.

Icon

An icon used for the window titlebar, taskbar, etc.

Window

Represents a window.

WindowAttributes

Attributes used when creating a window.

WindowButtons

WindowId

Identifier of a window. Unique for each window.

Enums

BadIcon

An error produced when using Icon::from_rgba with invalid arguments.

BadImage

An error produced when using CustomCursor::from_rgba with invalid arguments.

Cursor

See Window::set_cursor() for more details.

CursorGrabMode

The behavior of cursor grabbing.

CursorIcon

Describes the appearance of the (usually mouse) cursor icon.

Fullscreen

Fullscreen modes.

ImePurpose

Generic IME purposes for use in Window::set_ime_purpose.

ResizeDirection

Defines the orientation that a window resize will be performed.

Theme

The theme variant to use.

UserAttentionType

Platform-specific

WindowLevel

A window level groups windows with respect to their z-position.

Constants

MAX_CURSOR_SIZE

The maximum width and height for a cursor when using CustomCursor::from_rgba.