Struct winit::window::Window

pub struct Window { /* fields omitted */ }

Represents a window.

Example

use winit::{
    event::{Event, WindowEvent},
    event_loop::{ControlFlow, EventLoop},
    window::Window,
};

let mut event_loop = EventLoop::new();
let window = Window::new(&event_loop).unwrap();

event_loop.run(move |event, _, control_flow| {
    *control_flow = ControlFlow::Wait;

    match event {
        Event::WindowEvent {
            event: WindowEvent::CloseRequested,
            ..
        } => *control_flow = ControlFlow::Exit,
        _ => (),
    }
});

Implementations

impl Window

Base Window functions.

pub fn new<T: 'static>(
event_loop: &EventLoopWindowTarget<T>
) -> Result<Window, OsError>

Creates a new Window for platforms where this is appropriate.

This function is equivalent to WindowBuilder::new().build(event_loop).

Error should be very rare and only occur in case of permission denied, incompatible system, out of memory, etc.

Platform-specific behavior:

Web: The window is created but not inserted into the web page automatically. Please see the web platform module for more information.

pub fn id(&self) -> WindowId

Returns an identifier unique to the window.

pub fn scale_factor(&self) -> f64

Returns the scale factor that can be used to map logical pixels to physical pixels, and vice versa.

See the dpi module for more information.

Note that this value can change depending on user action (for example if the window is moved to another screen); as such, tracking WindowEvent::ScaleFactorChanged events is the most robust way to track the DPI you need to use to draw.

Platform-specific

X11: This respects Xft.dpi, and can be overridden using the WINIT_X11_SCALE_FACTOR environment variable.

Android: Always returns 1.0.

iOS: Can only be called on the main thread. Returns the underlying UIView's contentScaleFactor.

pub fn request_redraw(&self)

Emits a WindowEvent::RedrawRequested event in the associated event loop after all OS events have been processed by the event loop.

This is the strongly encouraged method of redrawing windows, as it can integrate with OS-requested redraws (e.g. when a window gets resized).

This function can cause RedrawRequested events to be emitted after Event::MainEventsCleared but before Event::NewEvents if called in the following circumstances:

While processing MainEventsCleared.

While processing a RedrawRequested event that was sent during MainEventsCleared or any directly subsequent RedrawRequested event.

Platform-specific

iOS: Can only be called on the main thread.

impl Window

Position and size functions.

pub fn inner_position(&self) -> Result<PhysicalPosition<i32>, NotSupportedError>

Returns the position of the top-left hand corner of the window's client area relative to the top-left hand corner of the desktop.

The same conditions that apply to outer_position apply to this method.

Platform-specific

iOS: Can only be called on the main thread. Returns the top left coordinates of the window's safe area in the screen space coordinate system.

Web: Returns the top-left coordinates relative to the viewport. Note: this returns the same value as outer_position.

pub fn outer_position(&self) -> Result<PhysicalPosition<i32>, NotSupportedError>

Returns the position of the top-left hand corner of the window relative to the top-left hand corner of the desktop.

Note that the top-left hand corner of the desktop is not necessarily the same as the screen. If the user uses a desktop with multiple monitors, the top-left hand corner of the desktop is the top-left hand corner of the monitor at the top-left of the desktop.

The coordinates can be negative if the top-left hand corner of the window is outside of the visible screen region.

Platform-specific

iOS: Can only be called on the main thread. Returns the top left coordinates of the window in the screen space coordinate system.

Web: Returns the top-left coordinates relative to the viewport.

pub fn set_outer_position<P: Into<Position>>(&self, position: P)

Modifies the position of the window.

See outer_position for more information about the coordinates. This automatically un-maximizes the window if it's maximized.

Platform-specific

iOS: Can only be called on the main thread. Sets the top left coordinates of the window in the screen space coordinate system.

Web: Sets the top-left coordinates relative to the viewport.

pub fn inner_size(&self) -> PhysicalSize<u32>

Returns the physical size of the window's client area.

The client area is the content of the window, excluding the title bar and borders.

Platform-specific

iOS: Can only be called on the main thread. Returns the PhysicalSize of the window's safe area in screen space coordinates.

Web: Returns the size of the canvas element.

pub fn set_inner_size<S: Into<Size>>(&self, size: S)

Modifies the inner size of the window.

See inner_size for more information about the values. This automatically un-maximizes the window if it's maximized.

Platform-specific

iOS: Unimplemented. Currently this panics, as it's not clear what set_inner_size would mean for iOS.

Web: Sets the size of the canvas element.

pub fn outer_size(&self) -> PhysicalSize<u32>

Returns the physical size of the entire window.

These dimensions include the title bar and borders. If you don't want that (and you usually don't), use inner_size instead.

Platform-specific

iOS: Can only be called on the main thread. Returns the PhysicalSize of the window in screen space coordinates.

Web: Returns the size of the canvas element. Note: this returns the same value as inner_size.

pub fn set_min_inner_size<S: Into<Size>>(&self, min_size: Option<S>)

Sets a minimum dimension size for the window.

Platform-specific

iOS: Has no effect.

Web: Has no effect.

pub fn set_max_inner_size<S: Into<Size>>(&self, max_size: Option<S>)

Sets a maximum dimension size for the window.

Platform-specific

iOS: Has no effect.

Web: Has no effect.

impl Window

Misc. attribute functions.

pub fn set_title(&self, title: &str)

Modifies the title of the window.

Platform-specific

Has no effect on iOS.

pub fn set_visible(&self, visible: bool)

Modifies the window's visibility.

If false, this will hide the window. If true, this will show the window.

Platform-specific

Android: Has no effect.

iOS: Can only be called on the main thread.

Web: Has no effect.

pub fn set_resizable(&self, resizable: bool)

Sets whether the window is resizable or not.

Note that making the window unresizable doesn't exempt you from handling Resized, as that event can still be triggered by DPI scaling, entering fullscreen mode, etc.

Platform-specific

This only has an effect on desktop platforms.

Due to a bug in XFCE, this has no effect on Xfwm.

Platform-specific

iOS: Has no effect.

Web: Has no effect.

pub fn set_minimized(&self, minimized: bool)

Sets the window to minimized or back

Platform-specific

iOS: Has no effect

pub fn set_maximized(&self, maximized: bool)

Sets the window to maximized or back.

Platform-specific

iOS: Has no effect.

Web: Has no effect.

pub fn set_fullscreen(&self, fullscreen: Option<Fullscreen>)

Sets the window to fullscreen or back.

Platform-specific

macOS: Fullscreen::Exclusive provides true exclusive mode with a video mode change. Caveat! macOS doesn't provide task switching (or spaces!) while in exclusive fullscreen mode. This mode should be used when a video mode change is desired, but for a better user experience, borderless fullscreen might be preferred.

Fullscreen::Borderless provides a borderless fullscreen window on a separate space. This is the idiomatic way for fullscreen games to work on macOS. See WindowExtMacOs::set_simple_fullscreen if separate spaces are not preferred.

The dock and the menu bar are always disabled in fullscreen mode.

iOS: Can only be called on the main thread.

Wayland: Does not support exclusive fullscreen mode.

Windows: Screen saver is disabled in fullscreen mode.

pub fn fullscreen(&self) -> Option<Fullscreen>

Gets the window's current fullscreen state.

Platform-specific

iOS: Can only be called on the main thread.

pub fn set_decorations(&self, decorations: bool)

Turn window decorations on or off.

Platform-specific

iOS: Can only be called on the main thread. Controls whether the status bar is hidden via setPrefersStatusBarHidden.

Web: Has no effect.

pub fn set_always_on_top(&self, always_on_top: bool)

Change whether or not the window will always be on top of other windows.

Platform-specific

iOS: Has no effect.

Web: Has no effect.

pub fn set_window_icon(&self, window_icon: Option<Icon>)

Sets the window icon. On Windows and X11, this is typically the small icon in the top-left corner of the titlebar.

Platform-specific

This only has an effect on Windows and X11.

On Windows, this sets ICON_SMALL. The base size for a window icon is 16x16, but it's recommended to account for screen scaling and pick a multiple of that, i.e. 32x32.

X11 has no universal guidelines for icon sizes, so you're at the whims of the WM. That said, it's usually in the same ballpark as on Windows.

pub fn set_ime_position<P: Into<Position>>(&self, position: P)

Sets location of IME candidate box in client area coordinates relative to the top left.

Platform-specific

iOS: Has no effect.

Web: Has no effect.

impl Window

Cursor functions.

pub fn set_cursor_icon(&self, cursor: CursorIcon)

Modifies the cursor icon of the window.

Platform-specific

iOS: Has no effect.

Android: Has no effect.

pub fn set_cursor_position<P: Into<Position>>(
&self,
position: P
) -> Result<(), ExternalError>

Changes the position of the cursor in window coordinates.

Platform-specific

iOS: Always returns an Err.

Web: Has no effect.

pub fn set_cursor_grab(&self, grab: bool) -> Result<(), ExternalError>

Grabs the cursor, preventing it from leaving the window.

Platform-specific

macOS: This presently merely locks the cursor in a fixed location, which looks visually awkward.

Wayland: This presently merely locks the cursor in a fixed location, which looks visually awkward.

Android: Has no effect.

iOS: Always returns an Err.

Web: Has no effect.

pub fn set_cursor_visible(&self, visible: bool)

Modifies the cursor's visibility.

If false, this will hide the cursor. If true, this will show the cursor.

Platform-specific

Windows: The cursor is only hidden within the confines of the window.

X11: The cursor is only hidden within the confines of the window.

Wayland: The cursor is only hidden within the confines of the window.

macOS: The cursor is hidden as long as the window has input focus, even if the cursor is outside of the window.

iOS: Has no effect.

Android: Has no effect.

impl Window

Monitor info functions.

pub fn current_monitor(&self) -> MonitorHandle

Returns the monitor on which the window currently resides

Platform-specific

iOS: Can only be called on the main thread.

pub fn available_monitors(&self) -> impl Iterator<Item = MonitorHandle>

Returns the list of all the monitors available on the system.

This is the same as EventLoop::available_monitors, and is provided for convenience.

Platform-specific

iOS: Can only be called on the main thread.

pub fn primary_monitor(&self) -> MonitorHandle

Returns the primary monitor of the system.

This is the same as EventLoop::primary_monitor, and is provided for convenience.

Platform-specific

iOS: Can only be called on the main thread.