Struct winit::window::WindowBuilder

pub struct WindowBuilder { /* private fields */ }

Object that allows building windows.

Implementations

impl WindowBuilder

pub fn new() -> Self

Initializes a new builder with default values.

pub fn with_inner_size<S: Into<Size>>(self, size: S) -> Self

Requests the window to be of specific dimensions.

If this is not set, some platform-specific dimensions will be used.

See Window::set_inner_size for details.

pub fn with_min_inner_size<S: Into<Size>>(self, min_size: S) -> Self

Sets the minimum dimensions a window can have.

If this is not set, the window will have no minimum dimensions (aside from reserved).

See Window::set_min_inner_size for details.

pub fn with_max_inner_size<S: Into<Size>>(self, max_size: S) -> Self

Sets the maximum dimensions a window can have.

If this is not set, the window will have no maximum or will be set to the primary monitorโs dimensions by the platform.

See Window::set_max_inner_size for details.

pub fn with_position<P: Into<Position>>(self, position: P) -> Self

Sets a desired initial position for the window.

If this is not set, some platform-specific position will be chosen.

See Window::set_outer_position for details.

Platform-specific

macOS: The top left corner position of the window content, the windowโs โinnerโ position. The window title bar will be placed above it. The window will be positioned such that it fits on screen, maintaining set inner_size if any. If you need to precisely position the top left corner of the whole window you have to use Window::set_outer_position after creating the window.

Windows: The top left corner position of the window title bar, the windowโs โouterโ position. There may be a small gap between this position and the window due to the specifics of the Window Manager.

X11: The top left corner of the window, the windowโs โouterโ position.

Others: Ignored.

pub fn with_resizable(self, resizable: bool) -> Self

Sets whether the window is resizable or not.

The default is true.

See Window::set_resizable for details.

pub fn with_title<T: Into<String>>(self, title: T) -> Self

Sets the initial title of the window in the title bar.

The default is "winit window".

See Window::set_title for details.

pub fn with_fullscreen(self, fullscreen: Option<Fullscreen>) -> Self

Sets whether the window should be put into fullscreen upon creation.

The default is None.

See Window::set_fullscreen for details.

pub fn with_maximized(self, maximized: bool) -> Self

Request that the window is maximized upon creation.

The default is false.

See Window::set_maximized for details.

pub fn with_visible(self, visible: bool) -> Self

Sets whether the window will be initially visible or hidden.

The default is to show the window.

See Window::set_visible for details.

pub fn with_transparent(self, transparent: bool) -> Self

Sets whether the background of the window should be transparent.

If this is true, writing colors with alpha values different than 1.0 will produce a transparent window.

The default is false.

pub fn transparent(&self) -> bool

Get whether the window will support transparency.

pub fn with_decorations(self, decorations: bool) -> Self

Sets whether the window should have a border, a title bar, etc.

The default is true.

See Window::set_decorations for details.

pub fn with_always_on_top(self, always_on_top: bool) -> Self

Sets whether or not the window will always be on top of other windows.

The default is false.

See Window::set_always_on_top for details.

pub fn with_window_icon(self, window_icon: Option<Icon>) -> Self

Sets the window icon.

The default is None.

See Window::set_window_icon for details.

pub fn build<T: 'static>( self, window_target: &EventLoopWindowTarget<T>, ) -> Result<Window, OsError>

Builds the window.

Possible causes of error include denied permission, incompatible system, and lack of memory.

Platform-specific

Web: The window is created but not inserted into the web page automatically. Please see the web platform module for more information.