Struct winit::window::WindowBuilder

pub struct WindowBuilder {
    pub window: WindowAttributes,
    // some fields omitted
}

Object that allows you to build windows.

Fields
window: WindowAttributes

The attributes to use to create the window.

Methods

impl WindowBuilder

pub fn new() -> Self

Initializes a new WindowBuilder with default values.

pub fn with_inner_size(self, size: LogicalSize) -> Self

Requests the window to be of specific dimensions.

See Window::set_inner_size for details.

pub fn with_min_inner_size(self, min_size: LogicalSize) -> Self

Sets a minimum dimension size for the window.

See Window::set_min_inner_size for details.

pub fn with_max_inner_size(self, max_size: LogicalSize) -> Self

Sets a maximum dimension size for the window.

See Window::set_max_inner_size for details.

pub fn with_resizable(self, resizable: bool) -> Self

Sets whether the window is resizable or not.

See Window::set_resizable for details.

pub fn with_title<T: Into<String>>(self, title: T) -> Self

Requests a specific title for the window.

See Window::set_title for details.

pub fn with_fullscreen(self, monitor: Option<Fullscreen>) -> Self

Sets the window fullscreen state.

See Window::set_fullscreen for details.

pub fn with_maximized(self, maximized: bool) -> Self

Requests maximized mode.

See Window::set_maximized for details.

pub fn with_visible(self, visible: bool) -> Self

Sets whether the window will be initially hidden or visible.

See Window::set_visible for details.

pub fn with_transparent(self, transparent: bool) -> Self

Sets whether the background of the window should be transparent.

pub fn with_decorations(self, decorations: bool) -> Self

Sets whether the window should have a border, a title bar, etc.

See Window::set_decorations for details.

pub fn with_always_on_top(self, always_on_top: bool) -> Self

Sets whether or not the window will always be on top of other windows.

See Window::set_always_on_top for details.

pub fn with_window_icon(self, window_icon: Option<Icon>) -> Self

Sets the window icon.

See Window::set_window_icon for details.

pub fn build<T: 'static>(
self,
window_target: &EventLoopWindowTarget<T>
) -> Result<Window, OsError>

Builds the window.

Possible causes of error include denied permission, incompatible system, and lack of memory.

Platform-specific behavior:

Web: The window is created but not inserted into the web page automatically. Please see the web platform module for more information.